chore: swap in new rewrite and remove old version
