+++ src/Uno.UI/UI/Xaml/Documents/UnicodeText.skia.cs
@@ -3,147 +3,79 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
-using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
 using Windows.Foundation;
 using Windows.UI.Text;
 using HarfBuzzSharp;
 using Microsoft.UI.Composition;
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Documents.TextFormatting;
 using Microsoft.UI.Xaml.Media;
 using SkiaSharp;
-using Uno.Extensions;
+using Uno.Buffers;
+using Uno.Disposables;
 using Uno.Foundation.Extensibility;
 using Uno.Foundation.Logging;
 using Uno.UI;
 using Uno.UI.Dispatching;
-using Uno.UI.Xaml.Media;
 using Buffer = HarfBuzzSharp.Buffer;
-using GlyphInfo = HarfBuzzSharp.GlyphInfo;
+using FontWeights = Microsoft.UI.Text.FontWeights;
 
 namespace Microsoft.UI.Xaml.Documents;
 
-// Excerpt from https://www.unicode.org/reports/tr9/tr9-6.html:
-// The Bidirectional Algorithm takes a stream of text as input, and proceeds in three main phases:
-//  * Separation of the input text into paragraphs. The rest of the algorithm affects only the text between paragraph separators.
-// 	* Resolution of the embedding levels of the text. In this phase, the directional character types, plus the explicit format codes, are used to produce resolved embedding levels.
-// 	* Reordering the text for display on a line-by-line basis using the resolved embedding levels, once the text has been broken into lines.
-// The algorithm only reorders text within a paragraph; characters in one paragraph have no effect on characters in a different paragraph. Paragraphs are divided by the Paragraph Separator or appropriate Newline Function (see Section 4.3, Directionality and Unicode Technical Report #13, “Unicode Newline Guidelines,” found on the CD-ROM or the up-to-date version on the Unicode web site on the handling of CR, LF, and CRLF). Paragraphs may also be determined by higher-level protocols: for example, the text in two different cells of a table will be in different paragraphs.
-
-// TODO: character spacing
-// TODO: what happens if text has no drawable glyphs but is not empty? Can this happen? The HarfBuzz docs imply that it can't
+// TODO tab stop handling with trimming etc
 internal readonly partial struct UnicodeText : IParsedText
 {
 	// Measured by hand from WinUI. Oddly enough, it doesn't depend on the font size.
 	private const float TabStopWidth = 48;
 	private const byte UBIDI_DEFAULT_LTR = 0xfe;
+	private const byte UBIDI_DEFAULT_RTL = 0xff;
 	private const int UBIDI_LTR = 0;
 	private const int UBIDI_RTL = 1;
 	private const string HorizontalEllipsis = "\u2026";
 
-	// A readonly snapshot of an Inline that is referenced by individual text runs after splitting. It's a class
-	// and not a struct because we don't want to copy the same Inline for each run.
-	private class ReadonlyInlineCopy
+	internal interface IFontCacheUpdateListener
 	{
-		public Inline? Inline { get; }
-		public int StartIndex { get; }
-		public int EndIndex { get; }
-		public string Text { get; }
-		public FlowDirection FlowDirection { get; }
-		public FontDetails FontDetails { get; }
-		public double FontSize { get; }
-		public FontWeight FontWeight { get; }
-		public FontStretch FontStretch { get; }
-		public FontStyle FontStyle { get; }
-		public Brush Foreground { get; }
-		public bool IsEllipsis { get; }
-
-		public ReadonlyInlineCopy(Inline inline, int startIndex, FlowDirection defaultFlowDirection, bool forceDefaultFlowDirection = false)
-		{
-			CI.Assert(inline is Run or LineBreak);
-			Inline = inline;
-			Text = inline.GetText();
-			Foreground = inline.Foreground;
-			FlowDirection = forceDefaultFlowDirection ? defaultFlowDirection : (inline as Run)?.FlowDirection ?? defaultFlowDirection;
-			FontDetails = inline.FontInfo;
-			FontSize = inline.FontSize;
-			FontWeight = inline.FontWeight;
-			FontStretch = inline.FontStretch;
-			FontStyle = inline.FontStyle;
-			StartIndex = startIndex;
-			EndIndex = startIndex + Text.Length;
-			IsEllipsis = false;
-		}
-
-		public static ReadonlyInlineCopy CreateEllipsis(BidiRun trimmedBidiRun, string trimmedText, int startIndex)
-			=> new(trimmedBidiRun, trimmedText, startIndex);
-		private ReadonlyInlineCopy(BidiRun trimmedBidiRun, string trimmedText, int startIndex)
-		{
-			Inline = null;
-			Text = trimmedText;
-			Foreground = trimmedBidiRun.inline.Foreground;
-			FlowDirection = trimmedBidiRun.rtl ? FlowDirection.RightToLeft : FlowDirection.LeftToRight;
-			FontDetails = trimmedBidiRun.fontDetails;
-			FontSize = trimmedBidiRun.inline.FontSize;
-			FontWeight = trimmedBidiRun.inline.FontWeight;
-			FontStretch = trimmedBidiRun.inline.FontStretch;
-			FontStyle = trimmedBidiRun.inline.FontStyle;
-			StartIndex = startIndex;
-			EndIndex = startIndex + Text.Length;
-			IsEllipsis = true;
-		}
+		void Invalidate();
 	}
 
-	// The same as HarfBuzz's GlyphPosition but with a float XAdvance instead of int
-	private readonly record struct UnoGlyphPosition(GlyphPosition GlyphPosition, float XAdvance);
+	private record struct Line(int start, int end, LinkedListNode<Cluster> clusterStart, LinkedListNode<Cluster> clusterLast, float width, float widthWithoutTrailingSpaces, float lineHeight, float baselineOffset, bool hasEllipsis = false);
 
-	// A BidiRun run split at line break boundaries
-	private readonly record struct BidiRun(ReadonlyInlineCopy inline, int startInInline, int endInInline, bool rtl, FontDetails fontDetails);
+	private record struct Glyph(GlyphPosition GlyphPosition, uint Codepoint);
 
-	// FontDetails might be different from inline.FontDetails because of font fallback
-	// glyphs are always ordered LTR even in RTL text
-	private readonly record struct ShapedLineBrokenBidiRun(ReadonlyInlineCopy Inline, int startInInline, int endInInline, (GlyphInfo info, UnoGlyphPosition position)[] glyphs, FontDetails fontDetails, bool rtl);
-	private record LayoutedGlyphDetails(GlyphInfo info, UnoGlyphPosition position, float xPosInRun, Cluster? cluster, LayoutedLineBrokenBidiRun parentRun)
+	private record struct Cluster(
+		int start,
+		int end,
+		LinkedListNode<Glyph> glyphStart,
+		LinkedListNode<Glyph> glyphLast,
+		FontDetails fontDetails,
+		float width,
+		bool containsOnlyWhitespace,
+		bool containsTab,
+		bool rtl,
+		int lineIndex,
+		int indexInLine)
 	{
-		public Cluster? cluster { get; set; } = cluster;
-	}
-
-	private record LayoutedLineBrokenBidiRun(ReadonlyInlineCopy inline, int startInInline, int endInInline, float xPosInLine, float width, LayoutedGlyphDetails[] glyphs, FontDetails fontDetails, bool rtl, LayoutedLine line, int indexInLine)
-	{
-		public float width { get; set; } = width;
-		public LayoutedLine line { get; set; } = line;
-
-		public int CompareTo(LayoutedLineBrokenBidiRun other)
+		public static Cluster Create(string _text, int indexStart, int indexEnd, LinkedListNode<Glyph> glyphsStart, LinkedListNode<Glyph> glyphsLast, FontDetails fontDetails)
 		{
-			return inline.StartIndex + startInInline - (other.inline.StartIndex + other.startInInline);
+			var clusterContainsTab = false;
+			var clusterContainsOnlyWhitespace = true;
+			for (int i = indexStart; i < indexEnd; i++)
+			{
+				clusterContainsTab |= _text[i] == '\t';
+				clusterContainsOnlyWhitespace &= char.IsWhiteSpace(_text[i]);
+			}
+
+			float clusterWidth = 0;
+			for (var glyphNode = glyphsStart; glyphNode != glyphsLast.Next; glyphNode = glyphNode.Next)
+			{
+				clusterWidth += AdvanceToPixels(glyphNode!.Value.GlyphPosition.XAdvance, fontDetails);
+			}
+
+			return new(indexStart, indexEnd, glyphsStart, glyphsLast, fontDetails, clusterWidth, clusterContainsOnlyWhitespace, clusterContainsTab, false, -1, -1);
 		}
-
-		public static bool operator <(LayoutedLineBrokenBidiRun left, LayoutedLineBrokenBidiRun right) =>
-			left.CompareTo(right) < 0;
-
-		public static bool operator >(LayoutedLineBrokenBidiRun left, LayoutedLineBrokenBidiRun right) =>
-			left.CompareTo(right) > 0;
-
-		public static bool operator <=(LayoutedLineBrokenBidiRun left, LayoutedLineBrokenBidiRun right) =>
-			left.CompareTo(right) <= 0;
-
-		public static bool operator >=(LayoutedLineBrokenBidiRun left, LayoutedLineBrokenBidiRun right) =>
-			left.CompareTo(right) >= 0;
 	}
-	// runs are always sorted LTR even in RTL text
-	// Each line must have at least one run except the very last line.
-	private record LayoutedLine(float lineHeight, float baselineOffset, int lineIndex, float xAlignmentOffset, float y, int startInText, int endInText, List<LayoutedLineBrokenBidiRun> runs);
-	private record Cluster(int sourceTextStart, int sourceTextEnd, LayoutedLineBrokenBidiRun layoutedRun, int glyphInRunIndexStart, int glyphInRunIndexEnd);
-
-	private static readonly Brush _blackBrush = new SolidColorBrush(Colors.Black);
-	private static readonly SKPaint _spareDrawPaint = new();
-	private static readonly SKPaint _spareSpellCheckPaint = new() { Color = SKColors.Red, Style = SKPaintStyle.Stroke, IsAntialias = true };
-
-	private static readonly Dictionary<int, HashSet<IFontCacheUpdateListener>> _codepointToListeners = new();
-	private static readonly Dictionary<string, HashSet<IFontCacheUpdateListener>> _fontFamilyToListeners = new();
 
 	private static readonly Lazy<ISpellCheckingService?> _spellCheckingService = new(() =>
 	{
@@ -158,26 +90,29 @@ internal readonly partial struct UnicodeText : IParsedText
 		}
 	});
 
-	private readonly Size _size;
-	private readonly TextAlignment _textAlignment;
-	private readonly bool _rtl;
-	private readonly List<ReadonlyInlineCopy> _inlines;
-	private readonly List<LayoutedLine> _lines;
-	private readonly Cluster[] _textIndexToGlyph;
-	private readonly Size _calculatedSize;
+	private static readonly Brush _blackBrush = new SolidColorBrush(Colors.Black);
+	private static readonly SKPaint _spareDrawPaint = new() { IsStroke = false, IsAntialias = true };
+	private static readonly SKPaint _spareSpellCheckPaint = new() { Color = SKColors.Red, Style = SKPaintStyle.Stroke, IsAntialias = true };
+	private static readonly Dictionary<int, HashSet<IFontCacheUpdateListener>> _codepointToListeners = new();
+	private static readonly Dictionary<string, HashSet<IFontCacheUpdateListener>> _fontFamilyToListeners = new();
 	private readonly string _text;
-	private readonly List<int> _wordBoundaries;
+	private readonly List<(float prefixSummedHeight, List<(float sumUntilAfterCluster, Cluster cluster)> prefixSummedWidths, Line line)> _xyTable;
+	private readonly List<(int start, int end, LinkedListNode<Cluster> cluster)> _indexToCluster;
+	private readonly TextAlignment _textAlignment;
 	private readonly FontDetails _defaultFontDetails;
+	private readonly List<Line> _lines;
+	private readonly float? _endingNewLineLineHeight;
+	private readonly bool _rtl;
+	private readonly List<(int start, int length, Hyperlink hyperlink)> _hyperlinkRanges;
+	private readonly List<int> _wordBoundaries;
+	private readonly List<LinkedListNode<Cluster>> _clustersInLogicalOrder;
+	private readonly LinkedList<Glyph> _glyphs;
+	private readonly List<(int end, FlowDirection direction)> _bidiBreaks;
+	private readonly List<(int end, Brush? foreground)> _runBreaks;
 	private readonly List<(int correctionStart, int correctionEnd)?>? _corrections;
+	private readonly Size _availableSize;
 
-	bool IParsedText.IsBaseDirectionRightToLeft => _rtl;
-
-	internal interface IFontCacheUpdateListener
-	{
-		void Invalidate();
-	}
-
-	internal UnicodeText(
+	internal unsafe UnicodeText(
 		Size availableSize,
 		Inline[] inlines, // only leaf nodes
 		FontDetails defaultFontDetails, // only used for a final empty line, otherwise the FontDetails are read from the inline
@@ -185,7 +120,7 @@ internal readonly partial struct UnicodeText : IParsedText
 		float lineHeight,
 		LineStackingStrategy lineStackingStrategy,
 		FlowDirection flowDirection,
-		TextAlignment? textAlignment, // null to determine from text. This will also infer the directionality of the text from the content
+		TextAlignment? textAlignment, // null to determine from text.
 		TextWrapping textWrapping,
 		TextTrimming textTrimming,
 		bool isSpellCheckEnabled,
@@ -193,431 +128,1216 @@ internal readonly partial struct UnicodeText : IParsedText
 		out Size calculatedSize)
 	{
 		CI.Assert(maxLines >= 0);
-		_size = availableSize;
-		_defaultFontDetails = defaultFontDetails;
 
-		string text;
-		if (textAlignment is null)
-		{
-			// TODO: can we make this cleaner instead of implicitly assuming that this is a code path coming from TextBox?
-			CI.Assert(inlines.Length == 1);
-			var inline = (Run)inlines[0];
-			var inlineText = inline.GetText();
-			if (inlineText.Length == 0)
-			{
-				flowDirection = inline.FlowDirection;
-			}
-			else
-			{
-				var firstInlineText = inlines[0].GetText();
-				using var _1 = ICU.CreateBiDiAndSetPara(firstInlineText, 0, firstInlineText.Length, UBIDI_DEFAULT_LTR, out var bidi);
-				ICU.GetMethod<ICU.ubidi_getLogicalRun>()(bidi, 0, out _, out var level);
-				CI.Assert(level is UBIDI_LTR or UBIDI_RTL);
-				flowDirection = level is UBIDI_RTL ? FlowDirection.RightToLeft : FlowDirection.LeftToRight;
-			}
-			textAlignment = flowDirection is FlowDirection.LeftToRight ? TextAlignment.Left : TextAlignment.Right;
-			var copy = new ReadonlyInlineCopy(inline, 0, flowDirection, true);
-			var length = copy.Text.Length;
-			_inlines = length == 0 ? [] : [copy];
-			text = copy.Text;
-		}
-		else
-		{
-			_inlines = new();
-			var lastEnd = 0;
-			var builder = new StringBuilder();
-			foreach (var inline in inlines)
-			{
-				var copy = new ReadonlyInlineCopy(inline, lastEnd, flowDirection);
-				var length = copy.Text.Length;
-				if (length != 0)
-				{
-					_inlines.Add(copy);
-				}
-				lastEnd = copy.EndIndex;
-				builder.Append(copy.Text);
-			}
-			text = builder.ToString();
-		}
+		var stringBuilder = new StringBuilder();
+		_runBreaks = new List<(int end, Brush? foreground)>();
+		var scriptBreaks = new List<int>();
+		var fontBreaks = new List<(int end, FontDetails fontDetails)>();
+		var lineOpportunityBreaks = new List<int>();
 
-		_rtl = flowDirection == FlowDirection.RightToLeft;
-
-		if (_inlines.Count == 0)
-		{
-			_lines = new();
-			calculatedSize = new Size(0, GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, defaultFontDetails, true, true).lineHeight);
-			_textIndexToGlyph = [];
-			_inlines = [];
-			_wordBoundaries = new();
-			_corrections = null;
-			_textAlignment = textAlignment.Value;
-			_text = "";
-			return;
-		}
-
-		var unlayoutedLines = SplitTextIntoLines(text, _rtl, _inlines, availableSize, textWrapping, maxLines, textTrimming, lineStackingStrategy, lineHeight, defaultFontDetails, fontListener);
-		var lines = LayoutLines(unlayoutedLines, textAlignment.Value, lineStackingStrategy, lineHeight, (float)availableSize.Width, defaultFontDetails);
-
-		_lines = maxLines == 0 || maxLines >= lines.Count ? lines : lines[..maxLines];
-		_text = text = text[.._lines[^1].endInText];
-
-		_textIndexToGlyph = new Cluster[_text.Length];
-		CreateSourceTextFromAndToGlyphMapping(_lines, _textIndexToGlyph);
-
-		_wordBoundaries = GetWords(text);
-		if (isSpellCheckEnabled && _spellCheckingService.Value is { } spellCheckingService)
-		{
-			_corrections = spellCheckingService.SpellCheck(_wordBoundaries, text);
-		}
-		else
-		{
-			_corrections = null;
-		}
-
-		var desiredHeight = _lines.Sum(l => l.lineHeight);
-		var desiredWidth = _lines.Max(l => l.runs.Sum(r => r.width));
-		_calculatedSize = calculatedSize = new Size(desiredWidth, desiredHeight);
-	}
-
-	/// <returns>The runs of each run are sorted according to the visual order.</returns>
-	private static List<(List<ShapedLineBrokenBidiRun> runs, int startInText, int endInText)> SplitTextIntoLines(
-		string text, bool rtl, List<ReadonlyInlineCopy> inlines, Size availableSize, TextWrapping textWrapping,
-		int maxLines, TextTrimming textTrimming, LineStackingStrategy lineStackingStrategy, float lineHeight,
-		FontDetails defaultFontDetails, IFontCacheUpdateListener fontListener)
-	{
-		var logicallyOrderedRuns = new List<BidiRun>();
-		var logicallyOrderedLineBreakingOpportunities = new List<(int indexInInline, ReadonlyInlineCopy inline)>();
 		foreach (var inline in inlines)
 		{
-			logicallyOrderedRuns.AddRange(SplitTextIntoLogicallyOrderedBidiRuns(inline, fontListener));
-			logicallyOrderedLineBreakingOpportunities.AddRange(GetLineBreakingOpportunities(inline.Text).Select(b => (b, inline)));
-		}
-
-		var lineBreakingOpporunities = SplitByLineBreakingOpportunities(logicallyOrderedRuns, logicallyOrderedLineBreakingOpportunities).ToList();
-		var lineBreakingIndices = lineBreakingOpporunities.Select(t => t.endIndexInLastRun + logicallyOrderedRuns[t.endRunIndex - 1].inline.StartIndex).ToList();
-
-		var lineWidthForLineBreaking = textWrapping == TextWrapping.NoWrap ? float.PositiveInfinity : (float)availableSize.Width;
-		var linesWithLogicallyOrderedRuns = ApplyLineBreaking(lineWidthForLineBreaking, logicallyOrderedRuns, lineBreakingOpporunities, rtl, textWrapping);
-
-		var lastLineNeedsEllipsis = RemoveOutOfViewLines(linesWithLogicallyOrderedRuns, maxLines, lineStackingStrategy, lineHeight, defaultFontDetails, availableSize);
-		if (textTrimming is TextTrimming.WordEllipsis)
-		{
-			ApplyWholeWordsTextTrimming(text, linesWithLogicallyOrderedRuns, lineBreakingIndices, (float)availableSize.Width, rtl, lastLineNeedsEllipsis);
-		}
-		else if (textTrimming is TextTrimming.CharacterEllipsis)
-		{
-			ApplyCharacterTextTrimming(text, linesWithLogicallyOrderedRuns, (float)availableSize.Width, rtl, lastLineNeedsEllipsis);
-		}
-		var linesWithBidiReordering = ApplyBidiReordering(rtl, linesWithLogicallyOrderedRuns, fontListener);
-
-		return linesWithBidiReordering;
-	}
-
-	private static void ApplyCharacterTextTrimming(string text, List<List<BidiRun>> linesWithLogicallyOrderedRuns, float availableWidth, bool rtl, bool lastLineNeedsEllipsis)
-	{
-		for (var lineIndex = 0; lineIndex < linesWithLogicallyOrderedRuns.Count; lineIndex++)
-		{
-			var line = linesWithLogicallyOrderedRuns[lineIndex];
-			if (line.Count == 0)
-			{
-				CI.Assert(lineIndex == linesWithLogicallyOrderedRuns.Count - 1);
-				return;
-			}
-
-			var lastRun = line[^1];
-			var fullLineWidthWithoutTrailingSpaces = MeasureContiguousRunSequence(0, line, 0, 0, line.Count, lastRun.endInInline, rtl);
-
-			if (fullLineWidthWithoutTrailingSpaces <= availableWidth && !(lastLineNeedsEllipsis && lineIndex == linesWithLogicallyOrderedRuns.Count - 1))
+			var inlineText = inline.GetText();
+			if (string.IsNullOrEmpty(inlineText))
 			{
 				continue;
 			}
 
-			var indexOfRunToTrimAt = 0;
-			(GlyphInfo info, UnoGlyphPosition position)[] ellipses = [];
-			var ellipsesWidth = 0f;
-			for (var endRunIndex = line.Count - 1; endRunIndex >= 0; endRunIndex--)
+			var inlineStart = stringBuilder.Length;
+			stringBuilder.Append(inlineText);
+
+			AppendBoundaries( /* Line */ 2, inlineText, inlineStart, lineOpportunityBreaks);
+
+			var currentFontDetails = inline.FontInfo;
+			int currentScript = 0;
+			for (var i = 0; i < inlineText.Length; i += char.IsSurrogate(inlineText, i) ? 2 : 1)
 			{
-				var testLastRun = line[endRunIndex];
-				var testEllipses = ShapeRun(HorizontalEllipsis, rtl, testLastRun.fontDetails, null, false);
-				var testEllipsesWidth = RunWidth(testEllipses, testLastRun.fontDetails);
-				var testLineWidthWithoutTrailingSpaces = MeasureContiguousRunSequence(0, line, 0, 0, endRunIndex + 1, testLastRun.endInInline, rtl);
-				if (testLineWidthWithoutTrailingSpaces + testEllipsesWidth <= availableWidth)
+				FontDetails newFontDetails;
+				var codepoint = char.ConvertToUtf32(inlineText, i);
+
+				var newScript = ICU.GetMethod<ICU.uscript_getScript>()(char.ConvertToUtf32(inlineText, i), out var errorCode);
+				ICU.CheckErrorCode<ICU.uscript_getScript>(errorCode);
+
+				if (newScript != currentScript)
 				{
-					indexOfRunToTrimAt = endRunIndex + 1;
-					break;
+					currentScript = newScript;
+					if (i != 0)
+					{
+						scriptBreaks.Add(inlineStart + i);
+					}
+				}
+
+				if (!inline.FontInfo.SKFont.ContainsGlyph(codepoint))
+				{
+					newFontDetails = GetFallbackFont(codepoint, (float)inline.FontSize, inline.FontWeight, inline.FontStretch, inline.FontStyle, fontListener) ?? inline.FontInfo;
 				}
 				else
 				{
-					(ellipses, ellipsesWidth) = (testEllipses, testEllipsesWidth);
+					newFontDetails = inline.FontInfo;
+				}
+
+				if (newFontDetails != currentFontDetails)
+				{
+					if (i != 0)
+					{
+						fontBreaks.Add((inlineStart + i, currentFontDetails));
+					}
+					currentFontDetails = newFontDetails;
 				}
 			}
 
-			if (ellipses.Length == 0)
-			{
-				ellipses = ShapeRun(HorizontalEllipsis, rtl, lastRun.fontDetails, null, false);
-				ellipsesWidth = RunWidth(ellipses, lastRun.fontDetails);
-			}
+			scriptBreaks.Add(inlineStart + inlineText.Length);
+			_runBreaks.Add((inlineStart + inlineText.Length, inline.Foreground));
+			fontBreaks.Add((inlineStart + inlineText.Length, currentFontDetails));
+		}
 
-			int indexAtTrim;
-			if (indexOfRunToTrimAt == line.Count)
+		_text = stringBuilder.ToString();
+		if (_text.Length == 0)
+		{
+			_lines = [];
+			_endingNewLineLineHeight = null;
+			_defaultFontDetails = defaultFontDetails;
+			_rtl = flowDirection is FlowDirection.RightToLeft;
+			_textAlignment = textAlignment ?? (flowDirection is FlowDirection.RightToLeft ? TextAlignment.Right : TextAlignment.Left);
+			_wordBoundaries = [];
+			var emptyHeight = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, defaultFontDetails, false, true).lineHeight;
+			calculatedSize = new Size(0, emptyHeight);
+			_availableSize = availableSize;
+			_xyTable = [];
+			_indexToCluster = [];
+			_hyperlinkRanges = [];
+			_clustersInLogicalOrder = [];
+			_glyphs = [];
+			_bidiBreaks = [];
+			return;
+		}
+		_hyperlinkRanges = new List<(int start, int length, Hyperlink hyperlink)>();
+
+		_bidiBreaks = new List<(int end, FlowDirection direction)>();
+		var embeddingLevels = ArrayPool<byte>.Shared.Rent(_text.Length);
+		using var embeddingLevelsDisposable = new DisposableStruct<byte[]>(static embeddingLevels => ArrayPool<byte>.Shared.Return(embeddingLevels), embeddingLevels);
+		for (int i = 0; i < inlines.Length; i++)
+		{
+			var (start, count) = i == 0 ? (0, _runBreaks[0].end) : (_runBreaks[i - 1].end, _runBreaks[i].end - _runBreaks[i - 1].end);
+			var direction = (inlines[i] as Run)?.FlowDirection ?? flowDirection;
+			var level = flowDirection is FlowDirection.LeftToRight
+				? (direction is FlowDirection.LeftToRight ? 0 : 1)
+				: (direction is FlowDirection.RightToLeft ? 1 : 2); // 2 and not 0 because embedding must increase nesting level when switching direction inside RTL paragraph
+			Array.Fill(embeddingLevels, (byte)level, start, count);
+
+			if (TryGetHyperLink(inlines[i]) is { } hyperLink)
 			{
-				indexAtTrim = lastRun.endInInline + lastRun.inline.StartIndex;
-				indexAtTrim -= TrailingCRLFCount(text, lastRun.inline.StartIndex + lastRun.startInInline, indexAtTrim);
-				indexAtTrim -= TrailingSpaceCount(text, lastRun.inline.StartIndex + lastRun.startInInline, indexAtTrim);
+				_hyperlinkRanges.Add((start, start + count, hyperLink));
+			}
+		}
+
+		using var _ = ICU.CreateBiDiAndSetPara(_text, 0, _text.Length, flowDirection is FlowDirection.RightToLeft ? UBIDI_DEFAULT_RTL : UBIDI_DEFAULT_LTR, out var bidi, embeddingLevels);
+		var runCount = ICU.GetMethod<ICU.ubidi_countRuns>()(bidi, out var countRunsErrorCode);
+		ICU.CheckErrorCode<ICU.ubidi_countRuns>(countRunsErrorCode);
+		_rtl = ICU.GetMethod<ICU.ubidi_getParaLevel>()(bidi) is UBIDI_RTL;
+		for (var bidiRunIndex = 0; bidiRunIndex < runCount; bidiRunIndex++)
+		{
+			var level = ICU.GetMethod<ICU.ubidi_getVisualRun>()(bidi, bidiRunIndex, out var logicalStart, out var length);
+			CI.Assert(level is UBIDI_LTR or UBIDI_RTL);
+			_bidiBreaks.Add((logicalStart + length, level is UBIDI_RTL ? FlowDirection.RightToLeft : FlowDirection.LeftToRight));
+
+			if (textAlignment is null && logicalStart == 0)
+			{
+				textAlignment = level is UBIDI_RTL ? TextAlignment.Right : TextAlignment.Left;
+			}
+		}
+
+		_glyphs = new LinkedList<Glyph>();
+		var clusterBreaks = new LinkedList<Cluster>();
+		foreach (var shapingRun in EnumerateShapingRuns(_runBreaks, scriptBreaks, _bidiBreaks, fontBreaks))
+		{
+			using var buffer = new Buffer();
+			buffer.AddUtf16(_text.AsSpan(shapingRun.start, shapingRun.end - shapingRun.start));
+			buffer.GuessSegmentProperties();
+			buffer.Direction = shapingRun.direction is FlowDirection.RightToLeft ? Direction.RightToLeft : Direction.LeftToRight;
+			shapingRun.fontDetails.Font.Shape(buffer);
+			var positions = buffer.GetGlyphPositionSpan();
+			var infos = buffer.GetGlyphInfoSpan();
+			var count = buffer.Length;
+
+			if (count == 0)
+			{
+				// Even though textRun is nonempty and fontDetails contains a font that can shape all the characters in it,
+				// Font.Shape may still decide to yield 0 glyphs.
+				_glyphs.AddLast(new Glyph { GlyphPosition = new GlyphPosition(), Codepoint = 0 });
 			}
 			else
 			{
-				var runToTrimAt = line[indexOfRunToTrimAt];
-				var endIndexInTrimRun = runToTrimAt.endInInline;
-				for (; endIndexInTrimRun > 0; endIndexInTrimRun--)
+				CI.Assert((buffer.Direction is Direction.LeftToRight && infos[0].Cluster == 0) || (buffer.Direction is Direction.RightToLeft && infos[^1].Cluster == 0));
+				if (buffer.Direction is Direction.LeftToRight)
 				{
-					var testLineWidthWithoutTrailingSpaces = MeasureContiguousRunSequence(0, line, 0, 0, indexOfRunToTrimAt + 1, endIndexInTrimRun, rtl);
-					if (testLineWidthWithoutTrailingSpaces + ellipsesWidth <= availableWidth)
+					for (var index = 0; index < infos.Length; index++)
 					{
-						line[indexOfRunToTrimAt] = runToTrimAt with { endInInline = endIndexInTrimRun };
-						break;
+						var info = infos[index];
+						var position = positions[index];
+						if (index > 0 && info.Cluster != infos[index - 1].Cluster)
+						{
+							clusterBreaks.AddLast(Cluster.Create(
+								_text,
+								clusterBreaks.Last?.Value.end ?? 0,
+								(int)(shapingRun.start + infos[index].Cluster),
+								clusterBreaks.Last?.Value.glyphLast?.Next ?? _glyphs.First!,
+								_glyphs.Last!,
+								shapingRun.fontDetails));
+						}
+						_glyphs.AddLast(new Glyph { GlyphPosition = position, Codepoint = info.Codepoint });
 					}
-				}
-
-				if (endIndexInTrimRun == 0)
-				{
-					indexOfRunToTrimAt--;
-					if (indexOfRunToTrimAt >= 0)
-					{
-						runToTrimAt = line[indexOfRunToTrimAt];
-						endIndexInTrimRun = runToTrimAt.endInInline;
-					}
-				}
-
-				if (indexOfRunToTrimAt >= 0)
-				{
-					indexAtTrim = runToTrimAt.inline.StartIndex + endIndexInTrimRun;
-					indexAtTrim -= TrailingCRLFCount(text, runToTrimAt.inline.StartIndex + runToTrimAt.startInInline, indexAtTrim);
-					indexAtTrim -= TrailingSpaceCount(text, runToTrimAt.inline.StartIndex + runToTrimAt.startInInline, indexAtTrim);
 				}
 				else
 				{
-					indexAtTrim = line[0].inline.StartIndex + line[0].startInInline + 1; // at least one character
-				}
-			}
-
-			var trimEnd = lineIndex == linesWithLogicallyOrderedRuns.Count - 1 ? text.Length : lastRun.inline.StartIndex + lastRun.endInInline;
-			var trimmedText = text[indexAtTrim..trimEnd];
-			var ellipsesRun = ReadonlyInlineCopy.CreateEllipsis(lastRun, trimmedText, indexAtTrim);
-			line.RemoveRange(indexOfRunToTrimAt + 1, line.Count - (indexOfRunToTrimAt + 1));
-			line.Add(lastRun with { endInInline = indexAtTrim - lastRun.inline.StartIndex });
-			line.Add(new BidiRun(ellipsesRun, 0, trimmedText.Length, lastRun.rtl, lastRun.fontDetails));
-		}
-	}
-
-	private static void ApplyWholeWordsTextTrimming(string text, List<List<BidiRun>> linesWithLogicallyOrderedRuns,
-		List<int> lineBreakingOpportunities, float availableWidth, bool rtl, bool lastLineNeedsEllipsis)
-	{
-		for (var lineIndex = 0; lineIndex < linesWithLogicallyOrderedRuns.Count; lineIndex++)
-		{
-			var line = linesWithLogicallyOrderedRuns[lineIndex];
-			if (line.Count == 0)
-			{
-				CI.Assert(lineIndex == linesWithLogicallyOrderedRuns.Count - 1);
-				return;
-			}
-
-			var fullLineWidthWithoutTrailingSpaces = MeasureContiguousRunSequence(0, line, 0, 0, line.Count, line[^1].endInInline, rtl);
-
-			if (fullLineWidthWithoutTrailingSpaces <= availableWidth && !(lastLineNeedsEllipsis && lineIndex == linesWithLogicallyOrderedRuns.Count - 1))
-			{
-				continue;
-			}
-
-			var lineBreakingOpportunityIndex = lineBreakingOpportunities.Count - 1;
-			var lineBreakingOpportunity = lineBreakingOpportunities[lineBreakingOpportunityIndex];
-
-			var foundTrimPoint = false;
-			for (var endRunIndex = line.Count - 1; !foundTrimPoint && endRunIndex >= 0; endRunIndex--)
-			{
-				while (lineBreakingOpportunityIndex > 0 && lineBreakingOpportunity > line[endRunIndex].endInInline + line[endRunIndex].inline.StartIndex)
-				{
-					lineBreakingOpportunityIndex--;
-					lineBreakingOpportunity = lineBreakingOpportunities[lineBreakingOpportunityIndex];
-				}
-
-				while (lineBreakingOpportunity > line[endRunIndex].startInInline + line[endRunIndex].inline.StartIndex)
-				{
-					lineBreakingOpportunity = lineBreakingOpportunities[lineBreakingOpportunityIndex];
-					var lastRun = line[endRunIndex];
-					var ellipses = ShapeRun(HorizontalEllipsis, rtl, lastRun.fontDetails, null, false);
-					var ellipsesWidth = RunWidth(ellipses, lastRun.fontDetails);
-					var (measuringEndRunIndex, endIndexInLastRun) = (endRunIndex + 1, lineBreakingOpportunity - lastRun.inline.StartIndex);
-					var testLineWidthWithoutTrailingSpaces = MeasureContiguousRunSequence(0, line, 0, 0, measuringEndRunIndex, endIndexInLastRun, rtl);
-					if (testLineWidthWithoutTrailingSpaces + ellipsesWidth <= availableWidth || lineBreakingOpportunityIndex == 0 || lineBreakingOpportunities[lineBreakingOpportunityIndex - 1] <= line[0].startInInline + line[0].inline.StartIndex)
+					for (var index = infos.Length - 1; index >= 0; index--)
 					{
-						var indexAtTrim = lineBreakingOpportunity;
-						indexAtTrim -= TrailingCRLFCount(text, lastRun.inline.StartIndex + lastRun.startInInline, indexAtTrim);
-						indexAtTrim -= TrailingSpaceCount(text, lastRun.inline.StartIndex + lastRun.startInInline, indexAtTrim);
-						var trimEnd = lineIndex == linesWithLogicallyOrderedRuns.Count - 1 ? text.Length : line[^1].inline.StartIndex + line[^1].endInInline;
-						var trimmedText = text[indexAtTrim..trimEnd];
-						var ellipsesRun = ReadonlyInlineCopy.CreateEllipsis(lastRun, trimmedText, indexAtTrim);
-						line.RemoveRange(endRunIndex, line.Count - endRunIndex);
-						line.Add(lastRun with { endInInline = indexAtTrim - lastRun.inline.StartIndex });
-						line.Add(new BidiRun(ellipsesRun, 0, trimmedText.Length, lastRun.rtl, lastRun.fontDetails));
-						foundTrimPoint = true;
-						break;
+						var info = infos[index];
+						var position = positions[index];
+						if (index < infos.Length - 1 && info.Cluster != infos[index + 1].Cluster)
+						{
+							clusterBreaks.AddLast(Cluster.Create(
+								_text,
+								clusterBreaks.Last?.Value.end ?? 0,
+								(int)(shapingRun.start + infos[index].Cluster),
+								clusterBreaks.Last?.Value.glyphLast?.Next ?? _glyphs.First!,
+								_glyphs.Last!,
+								shapingRun.fontDetails));
+						}
+						_glyphs.AddLast(new Glyph { GlyphPosition = position, Codepoint = info.Codepoint });
+					}
+				}
+
+				clusterBreaks.AddLast(Cluster.Create(
+					_text,
+					clusterBreaks.Last?.Value.end ?? 0,
+					shapingRun.end,
+					clusterBreaks.Last?.Value.glyphLast?.Next ?? _glyphs.First!,
+					_glyphs.Last!,
+					shapingRun.fontDetails));
+			}
+		}
+
+		var lines = new List<Line>();
+		{ // line breaking
+			float lineWidth = 0;
+			float lineWidthWithoutTrailingSpaces = 0;
+			int currentLineEnd = -1;
+			LinkedListNode<Cluster>? currentLineClusterLast = null;
+			FontDetails? maxHeightFontDetailsInCurrentLine = null;
+			// a "chunk" is a contiguous sequence of clusters with no line breaking opportunities that is being tested as a potential addition to the current line.
+			float chunkUnderTestWidth = 0;
+			float chunkUnderTestTrailingSpaceWidth = 0;
+			bool chunkUnderTestContainsOnlyWhitespace = true;
+			FontDetails? maxHeightFontDetailsInChunkUnderTest = null;
+			LinkedListNode<Cluster>? currentClusterBreak = clusterBreaks.First!;
+			for (var lineOpportunityBreakIndex = 0; lineOpportunityBreakIndex < lineOpportunityBreaks.Count; lineOpportunityBreakIndex++)
+			{
+				while (currentClusterBreak?.Value.end <= lineOpportunityBreaks[lineOpportunityBreakIndex])
+				{
+					var oldValues = (chunkUnderTestWidth, chunkUnderTestTrailingSpaceWidth, maxHeightFontDetailsInChunkUnderTest, chunkUnderTestContainsOnlyWhitespace);
+
+					var clusterWidth = currentClusterBreak.Value.containsTab
+						? ((int)((lineWidth + chunkUnderTestWidth) / TabStopWidth) + 1) * TabStopWidth - (lineWidth + chunkUnderTestWidth)
+						: currentClusterBreak.Value.width;
+
+					chunkUnderTestWidth += clusterWidth;
+					if (currentClusterBreak.Value is { containsOnlyWhitespace: true, containsTab: false })
+					{
+						chunkUnderTestTrailingSpaceWidth += clusterWidth;
 					}
 					else
 					{
-						lineBreakingOpportunityIndex--;
-						lineBreakingOpportunity = lineBreakingOpportunities[lineBreakingOpportunityIndex];
+						chunkUnderTestTrailingSpaceWidth = 0;
 					}
+
+					chunkUnderTestContainsOnlyWhitespace &= currentClusterBreak.Value is { containsOnlyWhitespace: true, containsTab: false };
+
+					if (maxHeightFontDetailsInChunkUnderTest is null || maxHeightFontDetailsInChunkUnderTest.LineHeight < currentClusterBreak.Value.fontDetails.LineHeight)
+					{
+						maxHeightFontDetailsInChunkUnderTest = currentClusterBreak.Value.fontDetails;
+					}
+
+					if (currentLineEnd == -1 && textWrapping is TextWrapping.Wrap && lineWidth + chunkUnderTestWidth - chunkUnderTestTrailingSpaceWidth > availableSize.Width)
+					{
+						float width;
+						float widthWithoutTrailingSpaces;
+						FontDetails fontDetails;
+						int end;
+						LinkedListNode<Cluster> clusterLast;
+						if (oldValues.maxHeightFontDetailsInChunkUnderTest is null) // this cluster is the only cluster in the line
+						{
+							width = chunkUnderTestWidth;
+							widthWithoutTrailingSpaces = chunkUnderTestWidth - chunkUnderTestTrailingSpaceWidth;
+							fontDetails = maxHeightFontDetailsInChunkUnderTest;
+							end = currentClusterBreak.Value.end;
+							clusterLast = currentClusterBreak;
+							if (currentClusterBreak.Value.containsTab)
+							{
+								// commit the final computed width of this tab stop
+								currentClusterBreak.Value = currentClusterBreak.Value with { width = clusterWidth };
+							}
+						}
+						else
+						{
+							width = oldValues.chunkUnderTestWidth;
+							widthWithoutTrailingSpaces = oldValues.chunkUnderTestWidth - oldValues.chunkUnderTestTrailingSpaceWidth;
+							fontDetails = oldValues.maxHeightFontDetailsInChunkUnderTest!;
+							end = currentClusterBreak.Value.start;
+							clusterLast = currentClusterBreak.Previous!;
+						}
+
+						var (h, b) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, fontDetails, lines.Count == 0, false);
+						lines.Add(new Line(lines.Count == 0 ? 0 : lines[^1].end, end, lines.Count == 0 ? clusterBreaks.First! : lines[^1].clusterLast.Next!, clusterLast, width, widthWithoutTrailingSpaces, h, b));
+						lineWidth = 0;
+						lineWidthWithoutTrailingSpaces = 0;
+						currentLineEnd = -1;
+						currentLineClusterLast = null;
+						maxHeightFontDetailsInCurrentLine = null;
+						chunkUnderTestWidth = 0;
+						chunkUnderTestTrailingSpaceWidth = 0;
+						chunkUnderTestContainsOnlyWhitespace = true;
+						maxHeightFontDetailsInChunkUnderTest = null;
+
+						if (oldValues.maxHeightFontDetailsInChunkUnderTest is null)
+						{
+							currentClusterBreak = currentClusterBreak.Next!;
+							continue;
+						}
+						lineOpportunityBreakIndex--;
+						break;
+					}
+
+					// cannot break line mid cluster, so only consider this a line break opportunity if the cluster ends with a line break opportunity
+					// A mandatory line break cannot occur mid cluster
+					// WinUI can always break after tabs, even in scenarios where ICU doesn't consider it a line break opportunity, so we follow suit
+					if (currentClusterBreak.Value.end == lineOpportunityBreaks[lineOpportunityBreakIndex] || currentClusterBreak.Value.containsTab)
+					{
+						if (lineWidth + chunkUnderTestWidth - chunkUnderTestTrailingSpaceWidth > availableSize.Width)
+						{
+							if (textWrapping is not TextWrapping.NoWrap && currentLineEnd != -1)
+							{
+								var (h, b) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, maxHeightFontDetailsInCurrentLine!, lines.Count == 0, false);
+								lines.Add(new Line(lines.Count == 0 ? 0 : lines[^1].end, currentLineEnd, lines.Count == 0 ? clusterBreaks.First! : lines[^1].clusterLast.Next!, currentLineClusterLast!, lineWidth, lineWidthWithoutTrailingSpaces, h, b));
+								lineWidth = 0;
+								lineWidthWithoutTrailingSpaces = 0;
+								maxHeightFontDetailsInCurrentLine = null;
+								currentLineEnd = -1;
+								currentLineClusterLast = null;
+								if (currentClusterBreak.Value.containsTab) // each "chunk" contains at most one tab, and always at the end
+								{
+									// recalculate the width of the tab
+									chunkUnderTestWidth -= clusterWidth;
+									clusterWidth = ((int)(chunkUnderTestWidth / TabStopWidth) + 1) * TabStopWidth - chunkUnderTestWidth;
+									chunkUnderTestWidth += clusterWidth;
+								}
+							}
+						}
+
+						currentLineEnd = currentClusterBreak.Value.end;
+						currentLineClusterLast = currentClusterBreak;
+						lineWidth += chunkUnderTestWidth;
+						if (!chunkUnderTestContainsOnlyWhitespace)
+						{
+							lineWidthWithoutTrailingSpaces = lineWidth - chunkUnderTestTrailingSpaceWidth;
+						}
+						if (maxHeightFontDetailsInCurrentLine is null || maxHeightFontDetailsInCurrentLine.LineHeight < maxHeightFontDetailsInChunkUnderTest.LineHeight)
+						{
+							maxHeightFontDetailsInCurrentLine = maxHeightFontDetailsInChunkUnderTest;
+						}
+
+						chunkUnderTestWidth = 0;
+						chunkUnderTestTrailingSpaceWidth = 0;
+						chunkUnderTestContainsOnlyWhitespace = true;
+						maxHeightFontDetailsInChunkUnderTest = null;
+
+						if (currentClusterBreak.Value.containsTab) // each "chunk" contains at most one tab, and always at the end
+						{
+							// commit the final computed width of this tab stop
+							currentClusterBreak.Value = currentClusterBreak.Value with { width = clusterWidth };
+						}
+
+						if (IsLineBreak(_text, currentClusterBreak.Value.end))
+						{
+							var (h, b) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, maxHeightFontDetailsInCurrentLine, lines.Count == 0, false);
+							lines.Add(new Line(lines.Count == 0 ? 0 : lines[^1].end, currentLineEnd, lines.Count == 0 ? clusterBreaks.First! : lines[^1].clusterLast.Next!, currentLineClusterLast, lineWidth, lineWidthWithoutTrailingSpaces, h, b));
+							lineWidth = 0;
+							lineWidthWithoutTrailingSpaces = 0;
+							maxHeightFontDetailsInCurrentLine = null;
+							currentLineEnd = -1;
+							currentLineClusterLast = null;
+						}
+
+						if (currentClusterBreak.Value.end == _text.Length && currentLineEnd != -1)
+						{
+							var (h, b) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, maxHeightFontDetailsInCurrentLine ?? defaultFontDetails, lines.Count == 0, true);
+							lines.Add(new Line(lines.Count == 0 ? 0 : lines[^1].end, currentLineEnd, lines.Count == 0 ? clusterBreaks.First! : lines[^1].clusterLast.Next!, currentLineClusterLast!, lineWidth, lineWidthWithoutTrailingSpaces, h, b));
+						}
+					}
+
+					currentClusterBreak = currentClusterBreak.Next;
 				}
 			}
 		}
-	}
 
-	private static bool RemoveOutOfViewLines(List<List<BidiRun>> linesWithLogicallyOrderedRuns, int maxLines, LineStackingStrategy lineStackingStrategy, float lineHeight, FontDetails defaultFontDetails, Size availableSize)
-	{
-		var resultingLineCount = linesWithLogicallyOrderedRuns.Count;
-		if (maxLines > 0)
+		var textEndsInLineBreak = IsLineBreak(_text, _text.Length);
+		float totalHeight = 0;
+		int nextTrimPointLookupStart = 0;
+		for (var lineIndex = 0; lineIndex < lines.Count; lineIndex++)
 		{
-			resultingLineCount = Math.Min(resultingLineCount, maxLines);
-		}
-		var totalHeight = 0f;
-		for (var lineIndex = 0; lineIndex < resultingLineCount; lineIndex++)
-		{
-			var line = linesWithLogicallyOrderedRuns[lineIndex];
-			float currentLineHeight;
-			if (line.Count == 0)
+			var line = lines[lineIndex];
+			totalHeight += line.lineHeight;
+			var nextLineHeight = lineIndex < lines.Count - 1
+				? lines[lineIndex + 1].lineHeight
+				: textEndsInLineBreak
+					? GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, defaultFontDetails, false, true).lineHeight
+					: 0;
+			var actualLineCount = lines.Count + (textEndsInLineBreak ? 1 : 0);
+			var isEarlyLastLine = (maxLines > 0 && maxLines < actualLineCount && lineIndex == maxLines - 1) || (lineIndex < actualLineCount - 1 && nextLineHeight + totalHeight > availableSize.Height);
+
+			var lineWidth = line.width;
+			LinkedListNode<Cluster> lastClusterIncludedInLine = line.clusterLast;
+			if (textTrimming is TextTrimming.CharacterEllipsis or TextTrimming.WordEllipsis && (line.widthWithoutTrailingSpaces > availableSize.Width || isEarlyLastLine))
 			{
-				// Only the last line can be empty. All other lines either have a line break character or actual content since you can't wrap to a new line without having any content on the initial line.
-				CI.Assert(lineIndex == linesWithLogicallyOrderedRuns.Count - 1 && lineIndex != 0);
-				(currentLineHeight, _) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, defaultFontDetails, false, true);
+				IEnumerable<LinkedListNode<Cluster>> possibleTrimPoints;
+				if (textTrimming is TextTrimming.WordEllipsis)
+				{
+					(possibleTrimPoints, nextTrimPointLookupStart) = EnumeratePossibleWordTrimmingBreaks(line, lineOpportunityBreaks, nextTrimPointLookupStart);
+				}
+				else
+				{
+					possibleTrimPoints = EnumeratePossibleCharacterTrimmingBreaks(line);
+				}
+
+				using var enumerator = possibleTrimPoints.GetEnumerator();
+				var hasMore = enumerator.MoveNext();
+				while (hasMore)
+				{
+					var trimPoint = enumerator.Current;
+					hasMore = enumerator.MoveNext();
+
+					// don't add ellipsis after white space, include the whitespace in the trimmed-out portion instead
+					while (trimPoint.Value is { containsOnlyWhitespace: true, containsTab: false } && trimPoint.Value.start > line.start)
+					{
+						trimPoint = trimPoint.Previous!;
+						if (enumerator.Current == trimPoint)
+						{
+							hasMore = enumerator.MoveNext();
+						}
+					}
+
+					while (lastClusterIncludedInLine != trimPoint)
+					{
+						lineWidth -= lastClusterIncludedInLine.Value.width;
+						lastClusterIncludedInLine = lastClusterIncludedInLine.Previous!;
+					}
+
+					using var buffer = new Buffer();
+					buffer.AddUtf16(HorizontalEllipsis);
+					buffer.GuessSegmentProperties();
+					var trimFontDetails = trimPoint.Value.fontDetails;
+					if (!trimFontDetails.SKFont.ContainsGlyph(HorizontalEllipsis[0]))
+					{
+						trimFontDetails = GetFallbackFont(HorizontalEllipsis[0], trimFontDetails.SKFontSize, FontWeights.Normal, FontStretch.Normal, FontStyle.Normal, fontListener) ?? trimFontDetails;
+					}
+					trimFontDetails.Font.Shape(buffer); // This can be cached and reused across trim points, but it's not expected to be a hotspot
+					var trimGlyphPositions = buffer.GetGlyphPositionSpan();
+					var trimGlyphInfos = buffer.GetGlyphInfoSpan();
+					float ellipsisWidth = 0;
+					foreach (var trimGlyphPosition in trimGlyphPositions)
+					{
+						ellipsisWidth += AdvanceToPixels(trimGlyphPosition.XAdvance, trimFontDetails);
+					}
+					if (lineWidth + ellipsisWidth <= availableSize.Width || !hasMore)
+					{
+						var clusterEnd = line.clusterLast.Next;
+						for (var clusterNode = trimPoint.Next; clusterNode is not null && clusterNode != clusterEnd;)
+						{
+							var next = clusterNode.Next;
+							clusterBreaks.Remove(clusterNode);
+							clusterNode = next;
+						}
+
+						var ellipsisGlyphList = new LinkedList<Glyph>();
+						for (var i = 0; i < trimGlyphInfos.Length; i++)
+						{
+							ellipsisGlyphList.AddLast(new Glyph(trimGlyphPositions[i], trimGlyphInfos[i].Codepoint));
+						}
+
+						var ellipsisCluster = new Cluster(
+							trimPoint.Value.end,
+							isEarlyLastLine ? _text.Length : line.end,
+							ellipsisGlyphList.First!,
+							ellipsisGlyphList.Last!,
+							trimFontDetails,
+							ellipsisWidth,
+							false,
+							false,
+							_rtl,
+							-1,
+							-1);
+						var ellipsisNode = trimPoint.List!.AddAfter(trimPoint, ellipsisCluster);
+						lines[lineIndex] = line = line with { clusterLast = ellipsisNode, widthWithoutTrailingSpaces = lineWidth + ellipsisWidth, width = lineWidth + ellipsisWidth, hasEllipsis = true, end = ellipsisCluster.end };
+						break;
+					}
+				}
 			}
-			else
+
+			if (isEarlyLastLine)
 			{
-				var fontDetailsWithMaxHeight = line.MaxBy(r => r.fontDetails.LineHeight).fontDetails;
-				(currentLineHeight, _) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, fontDetailsWithMaxHeight, lineIndex == 0, lineIndex == resultingLineCount - 1);
-			}
-			totalHeight += currentLineHeight;
-			if (totalHeight > availableSize.Height && lineIndex > 0)
-			{
-				resultingLineCount = lineIndex;
+				lines = lines[..(lineIndex + 1)];
 				break;
 			}
 		}
 
-		var linesRemoved = resultingLineCount < linesWithLogicallyOrderedRuns.Count;
-		linesWithLogicallyOrderedRuns.RemoveRange(resultingLineCount, linesWithLogicallyOrderedRuns.Count - resultingLineCount);
-		return linesRemoved;
-	}
+		_endingNewLineLineHeight = lines[^1].end == _text.Length && textEndsInLineBreak ? GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, defaultFontDetails, false, true).lineHeight : null;
+		totalHeight += _endingNewLineLineHeight ?? 0;
 
-	private static List<(List<ShapedLineBrokenBidiRun> runs, int startInText, int endInText)> ApplyBidiReordering(bool rtl, List<List<BidiRun>> linesWithLogicallyOrderedRuns, IFontCacheUpdateListener fontListener)
-	{
-		var shapedLines = new List<(List<ShapedLineBrokenBidiRun> runs, int startInText, int endInText)>();
-		for (var lineIndex = 0; lineIndex < linesWithLogicallyOrderedRuns.Count; lineIndex++)
+		float maxLineWidthWithoutTrailingSpaces = 0;
+		_indexToCluster = new List<(int start, int end, LinkedListNode<Cluster> cluster)>(clusterBreaks.Count);
+		for (var lineIndex = 0; lineIndex < lines.Count; lineIndex++)
 		{
-			var line = linesWithLogicallyOrderedRuns[lineIndex];
-			if (line.Count == 0)
+			var line = lines[lineIndex];
+			maxLineWidthWithoutTrailingSpaces = Math.Max(maxLineWidthWithoutTrailingSpaces, line.widthWithoutTrailingSpaces);
+			for (var node = line.clusterStart; ; node = node.Next!)
 			{
-				// Only the last line can be empty, otherwise it will have at least one piece of text or a line break.
-				CI.Assert(lineIndex == linesWithLogicallyOrderedRuns.Count - 1);
-				if (lineIndex == 0)
+				node.Value = node.Value with { lineIndex = lineIndex };
+				_indexToCluster.Add((node.Value.start, node.Value.end, node));
+				if (node == line.clusterLast)
 				{
-					shapedLines.Add((new(), 0, 0));
-				}
-				else
-				{
-					shapedLines.Add((new(), shapedLines[^1].endInText, shapedLines[^1].endInText));
-				}
-				continue;
-			}
-
-			var lineRuns = new Deque<ShapedLineBrokenBidiRun>();
-
-			foreach (var (inline, startInInline, endInInline) in GroupByInline(line))
-			{
-				if (inline.IsEllipsis)
-				{
-					var ellipsisRun = line[^1];
-					var ellipsisGlyphs = ShapeRun(HorizontalEllipsis, ellipsisRun.rtl, ellipsisRun.fontDetails, null, false);
-					var shapedEllipsisRun = new ShapedLineBrokenBidiRun(inline, ellipsisRun.startInInline, ellipsisRun.endInInline, ellipsisGlyphs, ellipsisRun.fontDetails, ellipsisRun.rtl);
-					if (rtl)
-					{
-						lineRuns.AddToFront(shapedEllipsisRun);
-					}
-					else
-					{
-						lineRuns.AddToBack(shapedEllipsisRun);
-					}
-					continue;
-				}
-
-				var sameInlineRuns = new List<ShapedLineBrokenBidiRun>();
-
-				using var _ = ICU.CreateBiDiAndSetPara(inline.Text, startInInline, endInInline, (byte)(inline.FlowDirection is FlowDirection.RightToLeft ? 1 : 0), out var bidi);
-
-				var runCount = ICU.GetMethod<ICU.ubidi_countRuns>()(bidi, out var countRunsErrorCode);
-				ICU.CheckErrorCode<ICU.ubidi_countRuns>(countRunsErrorCode);
-
-				for (var runIndex = 0; runIndex < runCount; runIndex++)
-				{
-					var level = ICU.GetMethod<ICU.ubidi_getVisualRun>()(bidi, runIndex, out var logicalStart, out var length);
-					CI.Assert(level is UBIDI_LTR or UBIDI_RTL);
-
-					var sameInlineRunsLengthBeforeFontSplitting = sameInlineRuns.Count;
-					var currentFontDetails = inline.FontDetails;
-					var currentFontSplitStart = logicalStart;
-					for (var i = logicalStart; i < logicalStart + length; i += char.IsSurrogate(inline.Text, startInInline + i) ? 2 : 1)
-					{
-						FontDetails newFontDetails;
-						if (char.ConvertToUtf32(inline.Text, startInInline + i) is var codepoint && !inline.FontDetails.SKFont.ContainsGlyph(codepoint))
-						{
-							newFontDetails = GetFallbackFont(codepoint, (float)inline.FontSize, inline.FontWeight, inline.FontStretch, inline.FontStyle, fontListener) ?? inline.FontDetails;
-						}
-						else
-						{
-							newFontDetails = inline.FontDetails;
-						}
-
-						if (newFontDetails != currentFontDetails)
-						{
-							if (i != logicalStart)
-							{
-								// the currentLineWidth parameter of ShapeRun is null and the tab width will be adjusted later in LayoutLines.
-								sameInlineRuns.Add(new ShapedLineBrokenBidiRun(inline, startInInline + currentFontSplitStart, startInInline + i, ShapeRun(inline.Text[(startInInline + currentFontSplitStart)..(startInInline + i)], level is UBIDI_RTL, currentFontDetails, null, ignoreTrailingSpaces: false), currentFontDetails, level is UBIDI_RTL));
-							}
-							currentFontDetails = newFontDetails;
-							currentFontSplitStart = i;
-						}
-					}
-					// the currentLineWidth parameter of ShapeRun is null and the tab width will be adjusted later in LayoutLines.
-					sameInlineRuns.Add(new ShapedLineBrokenBidiRun(inline, startInInline + currentFontSplitStart, startInInline + logicalStart + length, ShapeRun(inline.Text[(startInInline + currentFontSplitStart)..(startInInline + logicalStart + length)], level is UBIDI_RTL, currentFontDetails, null, ignoreTrailingSpaces: false), currentFontDetails, level is UBIDI_RTL));
-					// swap runs if rtl since we always process characters in a single bidi run in logical order
-					if (level is UBIDI_RTL)
-					{
-						for (var i = 0; i < (sameInlineRuns.Count - sameInlineRunsLengthBeforeFontSplitting) / 2; i++)
-						{
-							(sameInlineRuns[sameInlineRunsLengthBeforeFontSplitting + i], sameInlineRuns[^(1 + i)]) = (sameInlineRuns[^(1 + i)], sameInlineRuns[sameInlineRunsLengthBeforeFontSplitting + i]);
-						}
-					}
-				}
-				if (rtl)
-				{
-					for (int i = sameInlineRuns.Count - 1; i >= 0; i--)
-					{
-						lineRuns.AddToFront(sameInlineRuns[i]);
-					}
-				}
-				else
-				{
-					for (int i = 0; i < sameInlineRuns.Count; i++)
-					{
-						lineRuns.AddToBack(sameInlineRuns[i]);
-					}
+					break;
 				}
 			}
-			shapedLines.Add((lineRuns.ToList(), line[0].startInInline + line[0].inline.StartIndex, line[^1].endInInline + line[^1].inline.StartIndex));
 		}
 
-		return shapedLines;
+		_clustersInLogicalOrder = new(clusterBreaks.Count);
+		for (var node = clusterBreaks.First; ; node = node!.Next)
+		{
+			_clustersInLogicalOrder.Add(node!);
+			if (node == clusterBreaks.Last)
+			{
+				break;
+			}
+		}
+
+		for (var lineIndex = 0; lineIndex < lines.Count; lineIndex++)
+		{
+			var line = lines[lineIndex];
+			var ellipsisCluster = line.hasEllipsis ? line.clusterLast : null;
+			var limit = line.hasEllipsis ? line.clusterLast.Value.start : line.clusterLast.Value.end;
+			var lineBidi = ICU.GetMethod<ICU.ubidi_open>()();
+			ICU.GetMethod<ICU.ubidi_setLine>()(bidi, line.start, limit, lineBidi, out var setLineErrorCode);
+			ICU.CheckErrorCode<ICU.ubidi_setLine>(setLineErrorCode);
+			using var lineBidiDisposable = new DisposableStruct<IntPtr>(static bidi => ICU.GetMethod<ICU.ubidi_close>()(bidi), lineBidi);
+
+			var logicalToVisualMap = ArrayPool<int>.Shared.Rent(_text.Length);
+			using var logicalToVisualMapDisposable = new DisposableStruct<int[]>(static logicalToVisualMap => ArrayPool<int>.Shared.Return(logicalToVisualMap), logicalToVisualMap);
+
+			fixed (int* logicalToVisualMapPtr = logicalToVisualMap)
+			{
+				ICU.GetMethod<ICU.ubidi_getLogicalMap>()(lineBidi, (IntPtr)logicalToVisualMapPtr, out var getLogicalMapErrorCode);
+				ICU.CheckErrorCode<ICU.ubidi_getLogicalMap>(getLogicalMapErrorCode);
+			}
+
+			var levels = ICU.GetMethod<ICU.ubidi_getLevels>()(lineBidi, out var getLevelsErrorCode);
+			ICU.CheckErrorCode<ICU.ubidi_getLevels>(getLevelsErrorCode);
+			var levelsSpan = new Span<byte>(levels.ToPointer(), limit - line.start);
+
+			var clusterBeforeBegin = line.clusterStart.Previous;
+
+			var nodes = new List<LinkedListNode<Cluster>>();
+			for (var (clusterNode, end) = (line.clusterStart, (ellipsisCluster ?? line.clusterLast.Next)); clusterNode != end;)
+			{
+				clusterNode!.Value = clusterNode.Value with { rtl = levelsSpan[clusterNode.Value.start - line.start] % 2 == 1 };
+				var next = clusterNode.Next;
+				nodes.Add(clusterNode);
+				clusterBreaks.Remove(clusterNode);
+				clusterNode = next;
+			}
+
+			if (ellipsisCluster is not null)
+			{
+				clusterBreaks.Remove(ellipsisCluster);
+				ellipsisCluster.Value = ellipsisCluster.Value with { indexInLine = _rtl ? 0 : nodes.Count };
+			}
+
+			nodes.Sort((node1, node2) => logicalToVisualMap[node1.Value.start - line.start].CompareTo(logicalToVisualMap[node2.Value.start - line.start]));
+
+			for (var index = 0; index < nodes.Count; index++)
+			{
+				var clusterNode = nodes[index];
+				clusterNode.Value = clusterNode.Value with { indexInLine = index + (ellipsisCluster is not null && _rtl ? 1 : 0) };
+				var anchorNode = index == 0 ? clusterBeforeBegin : nodes[index - 1];
+				if (anchorNode is null)
+				{
+					clusterBreaks.AddFirst(clusterNode);
+				}
+				else
+				{
+					clusterBreaks.AddAfter(anchorNode, clusterNode);
+				}
+			}
+
+			var newFirstNode = nodes[0];
+			var newLastNode = nodes[^1];
+			if (ellipsisCluster is not null)
+			{
+				if (flowDirection is FlowDirection.LeftToRight)
+				{
+					clusterBreaks.AddAfter(nodes[^1], ellipsisCluster);
+					newLastNode = ellipsisCluster;
+				}
+				else
+				{
+					clusterBreaks.AddBefore(nodes[0], ellipsisCluster);
+					newFirstNode = ellipsisCluster;
+				}
+			}
+			lines[lineIndex] = line = line with { clusterStart = newFirstNode, clusterLast = newLastNode };
+		}
+
+		_xyTable = new List<(float prefixSummedHeight, List<(float sumUntilAfterCluster, Cluster cluster)> prefixSummedWidths, Line line)>(lines.Count);
+		float prefixSummedHeight = 0;
+		foreach (var line in lines)
+		{
+			prefixSummedHeight += line.lineHeight;
+
+			var prefixSummedWidths = new List<(float sumUntilAfterCluster, Cluster cluster)>();
+			float sumUntilAfterCluster = 0;
+			for (var (node, end) = (line.clusterStart, line.clusterLast.Next); node != end; node = node.Next!)
+			{
+				sumUntilAfterCluster += node.Value.width;
+				prefixSummedWidths.Add((sumUntilAfterCluster, node.Value));
+			}
+
+			_xyTable.Add((prefixSummedHeight, prefixSummedWidths, line));
+		}
+
+		_lines = lines;
+		_defaultFontDetails = defaultFontDetails;
+		_textAlignment = textAlignment!.Value;
+		_wordBoundaries = GetWords(_text);
+		_corrections = isSpellCheckEnabled ? _spellCheckingService.Value?.SpellCheck(_wordBoundaries, _text) : null;
+		calculatedSize = new Size(maxLineWidthWithoutTrailingSpaces, totalHeight);
+		_availableSize = availableSize;
+	}
+
+	private static IEnumerable<LinkedListNode<Cluster>> EnumeratePossibleCharacterTrimmingBreaks(Line line)
+	{
+		for (var i = line.clusterLast; ; i = i.Previous!)
+		{
+			if (i == line.clusterStart)
+			{
+				yield break;
+			}
+
+			yield return i;
+		}
+	}
+
+	private static (IEnumerable<LinkedListNode<Cluster>> possibleTrimPoints, int nextLookupStart) EnumeratePossibleWordTrimmingBreaks(Line line, List<int> lineBreakOpportunities, int lineBreakOpportunitieslookupStart)
+	{
+		var possibleTrimPoints = new Stack<LinkedListNode<Cluster>>();
+		var currectCluster = line.clusterStart;
+		for (var currentlineBreakOpportunity = lineBreakOpportunities[lineBreakOpportunitieslookupStart];
+			 lineBreakOpportunitieslookupStart < lineBreakOpportunities.Count && currentlineBreakOpportunity <= line.end;
+			 lineBreakOpportunitieslookupStart++)
+		{
+			currentlineBreakOpportunity = lineBreakOpportunities[lineBreakOpportunitieslookupStart];
+			while (currectCluster.Value.end < currentlineBreakOpportunity)
+			{
+				currectCluster = currectCluster.Next!;
+			}
+
+			if (currectCluster.Value.end == currentlineBreakOpportunity)
+			{
+				possibleTrimPoints.Push(currectCluster);
+			}
+		}
+
+		return (possibleTrimPoints, lineBreakOpportunitieslookupStart);
+	}
+
+	private static List<(int start, int end, FontDetails fontDetails, FlowDirection direction)> EnumerateShapingRuns(
+		List<(int end, Brush? foreground)> runBreaks,
+		List<int> scriptBreaks,
+		List<(int end, FlowDirection direction)> bidiBreaks,
+		List<(int end, FontDetails fontDetails)> fontBreaks)
+	{
+		var shapingRuns = new List<(int start, int end, FontDetails fontDetails, FlowDirection direction)>();
+
+		int currentRunBreakIndex = 0;
+		int currentScriptBreakIndex = 0;
+		int currentBidiBreakIndex = 0;
+		int currentFontBreakIndex = 0;
+
+		var start = 0;
+
+		while (currentRunBreakIndex < runBreaks.Count)
+		{
+			var nextRunBreak = runBreaks[currentRunBreakIndex];
+			var nextScriptBreak = scriptBreaks[currentScriptBreakIndex];
+			var nextBidiBreak = bidiBreaks[currentBidiBreakIndex].end;
+			var nextFontBreak = fontBreaks[currentFontBreakIndex].end;
+
+			var nextBreak = Math.Min(Math.Min(nextRunBreak.end, nextScriptBreak), Math.Min(nextBidiBreak, nextFontBreak));
+
+			if (nextBreak > start)
+			{
+				shapingRuns.Add((start, nextBreak, fontBreaks[currentFontBreakIndex].fontDetails, bidiBreaks[currentBidiBreakIndex].direction));
+				start = nextBreak;
+			}
+
+			if (nextBreak == nextRunBreak.end)
+			{
+				currentRunBreakIndex++;
+			}
+			if (nextBreak == nextScriptBreak)
+			{
+				currentScriptBreakIndex++;
+			}
+			if (nextBreak == nextBidiBreak)
+			{
+				currentBidiBreakIndex++;
+			}
+			if (nextBreak == nextFontBreak)
+			{
+				currentFontBreakIndex++;
+			}
+		}
+
+		return shapingRuns;
+	}
+
+	public void Draw(in Visual.PaintingSession session,
+		(int index, CompositionBrush brush, float thickness)? caret, // null to skip drawing a caret
+		IEnumerable<TextHighlighter> highlighters)
+	{
+		var highlighterSlicer = new RangeSlicer<(CompositionBrush? background, Brush foreground)>(0, _text.Length);
+		foreach (var highlighter in highlighters)
+		{
+			foreach (var range in highlighter.Ranges)
+			{
+				if (range.Length != 0 && range.StartIndex < _text.Length && _text.Length > 0)
+				{
+					highlighterSlicer.Mark(
+						Math.Min(range.StartIndex, _text.Length),
+						Math.Min(range.StartIndex + range.Length, _text.Length),
+						(highlighter.Background?.GetOrCreateCompositionBrush(Compositor.GetSharedCompositor()),
+							highlighter.Foreground ?? _blackBrush));
+				}
+			}
+		}
+
+		Dictionary<SKColor, Dictionary<SKFont, (List<ushort> glyphs, List<SKPoint> positions)>> _colorToFontToGlyphs = new();
+		List<(SKPath path, float strokeThickness)> spellCheckUnderlines = new();
+
+		SKRect? caretRect = default;
+
+		var runBreakIndex = 0;
+		var wordBoundariesIndex = 0;
+		var highlighterSlices = highlighterSlicer.GetSegments();
+		var highlighterIndex = 0;
+		for (var clusterIndex = 0; clusterIndex < _clustersInLogicalOrder.Count; clusterIndex++)
+		{
+			var cluster = _clustersInLogicalOrder[clusterIndex];
+			while (highlighterSlices[highlighterIndex].End <= cluster.Value.start)
+			{
+				highlighterIndex++;
+			}
+
+			var highlighter = highlighterSlices[highlighterIndex];
+
+			while (_runBreaks[runBreakIndex].end <= cluster.Value.start)
+			{
+				runBreakIndex++;
+			}
+
+			while (_wordBoundaries[wordBoundariesIndex] <= cluster.Value.start)
+			{
+				wordBoundariesIndex++;
+			}
+
+			var lineIndex = cluster.Value.lineIndex;
+			var line = _lines[lineIndex];
+			var y = _xyTable[lineIndex].prefixSummedHeight - line.lineHeight;
+			var unalignedX = cluster.Value.indexInLine == 0
+				? 0
+				: _xyTable[lineIndex].prefixSummedWidths[cluster.Value.indexInLine - 1].sumUntilAfterCluster;
+			var alignmentOffset = GetAlignmentOffsetForLine(line);
+			var positionAcc = new SKPoint(unalignedX + alignmentOffset, y + line.baselineOffset);
+			var fontDetails = cluster.Value.fontDetails;
+
+			if (!cluster.Value.containsTab)
+			{
+				var color = BrushToColor(highlighter.Value.foreground is { } h ? h : _runBreaks[runBreakIndex].foreground, session.Opacity);
+				if (!_colorToFontToGlyphs.TryGetValue(color, out var fontToGlyphs))
+				{
+					_colorToFontToGlyphs[color] = fontToGlyphs = new Dictionary<SKFont, (List<ushort> glyphs, List<SKPoint> positions)>();
+				}
+				if (!fontToGlyphs.TryGetValue(fontDetails.SKFont, out var glyphsAndPositions))
+				{
+					fontToGlyphs[fontDetails.SKFont] = glyphsAndPositions = (new List<ushort>(), new List<SKPoint>());
+				}
+				var glyphs = glyphsAndPositions.glyphs;
+				var positions = glyphsAndPositions.positions;
+
+				for (var glyphNode = cluster.Value.glyphStart; ; glyphNode = glyphNode.Next!)
+				{
+					var glyph = glyphNode.Value;
+					glyphs.Add((ushort)glyph.Codepoint);
+					positions.Add(new SKPoint(positionAcc.X + AdvanceToPixels(glyph.GlyphPosition.XOffset, fontDetails),
+						positionAcc.Y + AdvanceToPixels(glyph.GlyphPosition.YOffset, fontDetails)));
+					positionAcc.X += AdvanceToPixels(glyph.GlyphPosition.XAdvance, fontDetails);
+					if (cluster.Value.glyphLast == glyphNode)
+					{
+						break;
+					}
+				}
+			}
+
+			var backgroundRect = new SKRect(unalignedX + alignmentOffset, y, unalignedX + alignmentOffset + cluster.Value.width, y + line.lineHeight);
+			highlighter.Value.background?.Paint(session.Canvas, session.Opacity, backgroundRect);
+
+			if (_corrections?[wordBoundariesIndex] is { } correction)
+			{
+				var correctionIndexBase = wordBoundariesIndex == 0 ? 0 : _wordBoundaries[wordBoundariesIndex - 1];
+				if (correctionIndexBase + correction.correctionStart <= cluster.Value.start && correctionIndexBase + correction.correctionEnd >= cluster.Value.end)
+				{
+					var fontSize = fontDetails.SKFontSize;
+					var scale = fontSize / 12.0f;
+					var step = 4 * scale;
+					var amplitude = 2 * scale;
+					var yOffset = 2 * scale;
+
+					var p = new SKPath();
+					var underlineY = y + line.baselineOffset + yOffset;
+					var underlineLeftX = unalignedX + alignmentOffset;
+					var underlineRightX = underlineLeftX + cluster.Value.width;
+					p.MoveTo(underlineLeftX, underlineY);
+					var x = underlineLeftX;
+					var up = true;
+					while (x + step < underlineRightX)
+					{
+						x += step;
+						var yWave = underlineY + (up ? -amplitude : amplitude);
+						p.LineTo(x, yWave);
+						up = !up;
+					}
+					p.LineTo(underlineRightX, underlineY);
+
+					spellCheckUnderlines.Add((p, scale));
+				}
+			}
+
+			if (caret is var (caretIndex, _, caretThickness))
+			{
+				if (caretIndex >= cluster.Value.start && caretIndex < cluster.Value.end)
+				{
+					caretRect = cluster.Value.rtl
+						? new SKRect(cluster.Value.width + alignmentOffset + unalignedX - caretThickness, y, cluster.Value.width + alignmentOffset + unalignedX, y + line.lineHeight)
+						: new SKRect(alignmentOffset + unalignedX, y, alignmentOffset + unalignedX + caretThickness, y + line.lineHeight);
+				}
+				else if (_endingNewLineLineHeight is null && caretIndex >= cluster.Value.start && clusterIndex == _clustersInLogicalOrder.Count - 1)
+				{
+					caretRect = cluster.Value.rtl
+						? new SKRect(alignmentOffset + unalignedX - caretThickness, y, alignmentOffset + unalignedX, y + line.lineHeight)
+						: new SKRect(cluster.Value.width + alignmentOffset + unalignedX, y, cluster.Value.width + alignmentOffset + unalignedX + caretThickness, y + line.lineHeight);
+				}
+			}
+		}
+
+		// This would probably be more efficient with SKCanvas::DrawGlyphs, but it isn't exposed in SkiaSharp
+		using var textBlobBuilder = new SKTextBlobBuilder();
+		foreach (var (color, fontToGlyphs) in _colorToFontToGlyphs)
+		{
+			_spareDrawPaint.Color = color;
+			foreach (var (font, (glyphs, positions)) in fontToGlyphs)
+			{
+				textBlobBuilder.AddPositionedRun(CollectionsMarshal.AsSpan(glyphs), font, CollectionsMarshal.AsSpan(positions));
+				session.Canvas.DrawText(textBlobBuilder.Build(), 0, 0, _spareDrawPaint); // SKTextBlobBuilder::Build resets the builder
+			}
+		}
+
+		foreach (var (path, strokeThickness) in spellCheckUnderlines)
+		{
+			_spareSpellCheckPaint.StrokeWidth = strokeThickness;
+			session.Canvas.DrawPath(path, _spareSpellCheckPaint);
+		}
+
+		if (caretRect is null && caret?.index == _text.Length) // ending new line or empty text
+		{
+			var alignmentOffset = GetAlignmentOffsetForLine(null);
+			var top = _text.Length == 0 ? 0 : _xyTable[^1].prefixSummedHeight;
+			caretRect = _rtl
+				? new SKRect(alignmentOffset - caret.Value.thickness, top, alignmentOffset, top + _defaultFontDetails.LineHeight)
+				: new SKRect(alignmentOffset, top, alignmentOffset + caret.Value.thickness, top + _defaultFontDetails.LineHeight);
+		}
+
+		if (caretRect is not null)
+		{
+			caret!.Value.brush.Paint(session.Canvas, session.Opacity, caretRect.Value);
+		}
+	}
+
+	public (int replaceIndexStart, int replaceIndexEnd, List<string> suggestions)? GetSpellCheckSuggestions(int correctionStart, int correctionEnd)
+	{
+		if (_spellCheckingService.Value is { } spellCheckingService)
+		{
+			return spellCheckingService.GetSpellCheckSuggestions(_text, _wordBoundaries, correctionStart, correctionEnd);
+		}
+		else
+		{
+			return null;
+		}
+	}
+
+	private static SKColor BrushToColor(Brush? brush, float opacity)
+	{
+		var color = SKColors.Black;
+		if (brush is SolidColorBrush scb)
+		{
+			var scbColor = scb.Color;
+			color = new SKColor(
+				red: scbColor.R,
+				green: scbColor.G,
+				blue: scbColor.B,
+				alpha: (byte)(scbColor.A * scb.Opacity * opacity));
+		}
+		else if (brush is GradientBrush gb)
+		{
+			var gbColor = gb.FallbackColorWithOpacity;
+			color = new SKColor(
+				red: gbColor.R,
+				green: gbColor.G,
+				blue: gbColor.B,
+				alpha: (byte)(gbColor.A * opacity));
+		}
+		else if (brush is XamlCompositionBrushBase xcbb)
+		{
+			var gbColor = xcbb.FallbackColorWithOpacity;
+			color = new SKColor(
+				red: gbColor.R,
+				green: gbColor.G,
+				blue: gbColor.B,
+				alpha: (byte)(gbColor.A * opacity));
+		}
+
+		return color;
+	}
+
+
+	private static Hyperlink? TryGetHyperLink(Inline inline)
+	{
+		DependencyObject? parent = inline;
+		while (parent is TextElement textElement)
+		{
+			if (parent is Hyperlink hyperlink)
+			{
+				return hyperlink;
+			}
+			parent = textElement.GetParent() as DependencyObject;
+		}
+
+		return null;
+	}
+
+	public Rect GetRectForIndex(int index)
+	{
+		index = Math.Min(index, _text.Length);
+
+		if (index == 0)
+		{
+			double alignmentOffset = string.IsNullOrEmpty(_text) ? GetAlignmentOffsetForLine(null) : GetAlignmentOffsetForLine(_lines[0]);
+			return new Rect(alignmentOffset, 0, 0, _defaultFontDetails.LineHeight);
+		}
+
+		if (index == _text.Length)
+		{
+			var (alignmentOffset, lineWidth, height, y) = _endingNewLineLineHeight is { } endingNewLineLineHeight
+				? (GetAlignmentOffsetForLine(null), 0, endingNewLineLineHeight, _xyTable[^1].prefixSummedHeight)
+				: (GetAlignmentOffsetForLine(_lines[^1]), _lines[^1].width, _lines[^1].lineHeight, _xyTable[^1].prefixSummedHeight - _lines[^1].lineHeight);
+			return _rtl ?
+				new Rect(alignmentOffset, y, 0, height) :
+				new Rect(alignmentOffset + lineWidth, y, 0, height);
+		}
+		else
+		{
+			var clusterIndex = _indexToCluster.BinarySearch(
+				(index, index, null!),
+				Comparer<(int start, int end, LinkedListNode<Cluster> cluster)>.Create(static (a, b) => a.start.CompareTo(b.start)));
+
+			if (clusterIndex < 0)
+			{
+				clusterIndex = ~clusterIndex - 1;
+			}
+
+			var cluster = _indexToCluster[clusterIndex].cluster;
+			var lineIndex = cluster.Value.lineIndex;
+			var line = _lines[lineIndex];
+			var indexInLine = cluster.Value.indexInLine;
+
+			var alignmentOffset = GetAlignmentOffsetForLine(line);
+			var y = lineIndex == 0 ? 0 : _xyTable[lineIndex].prefixSummedHeight - _xyTable[lineIndex - 1].prefixSummedHeight;
+			var unalignedX = indexInLine == 0 ? (_rtl ? line.width : 0) : _xyTable[lineIndex].prefixSummedWidths[indexInLine - 1].sumUntilAfterCluster;
+			return new Rect(alignmentOffset + unalignedX, y, cluster.Value.width, line.lineHeight);
+		}
+	}
+
+	public int GetIndexAt(Point p, bool ignoreEndingNewLine, bool extendedSelection)
+	{
+		if (_text.Length == 0)
+		{
+			return extendedSelection ? 0 : -1;
+		}
+
+		if (p.Y < 0)
+		{
+			return extendedSelection ? 0 : -1;
+		}
+
+		if (p.Y >= _xyTable[^1].prefixSummedHeight + (_endingNewLineLineHeight ?? 0))
+		{
+			if (!extendedSelection)
+			{
+				return -1;
+			}
+			if (_rtl && p.X > GetAlignmentOffsetForLine(_lines[^1]) + _lines[^1].width || !_rtl && p.X < GetAlignmentOffsetForLine(_lines[^1]))
+			{
+				// corner case: bottom left (or bottom right if rtl) of the box, we can either go to the beginning or the end.
+				// we match winui and go to the beginning.
+
+				return 0;
+			}
+			else
+			{
+				return _text.Length - (ignoreEndingNewLine ? TrailingCRLFInLine(_endingNewLineLineHeight is null ? _lines[^1] : null) : 0);
+			}
+		}
+
+		if (p.Y >= _xyTable[^1].prefixSummedHeight)
+		{
+			return extendedSelection ? _lines[^1].end - (ignoreEndingNewLine ? TrailingCRLFInLine(_lines[^1]) : 0) : -1;
+		}
+
+		var lineIndex = _xyTable.BinarySearch(
+			((float)p.Y, null!, default),
+			Comparer<(float prefixSummedHeight, List<(float sumUntilAfterCluster, Cluster cluster)> prefixSummedWidths, Line line)>.Create(
+				static (a, b) => a.prefixSummedHeight.CompareTo(b.prefixSummedHeight)));
+
+		if (lineIndex < 0)
+		{
+			lineIndex = ~lineIndex;
+		}
+
+		var line = _xyTable[lineIndex].line;
+		var alignmentOffset = GetAlignmentOffsetForLine(line);
+
+		if (p.X < alignmentOffset)
+		{
+			return extendedSelection
+				? (_rtl ? line.end - (ignoreEndingNewLine ? TrailingCRLFInLine(line) : 0) : line.start)
+				: -1;
+		}
+
+		if (p.X >= alignmentOffset + line.width)
+		{
+			return extendedSelection
+				? (_rtl ? line.start : line.end - (ignoreEndingNewLine ? TrailingCRLFInLine(line) : 0))
+				: -1;
+		}
+
+		var prefixSummedWidths = _xyTable[lineIndex].prefixSummedWidths;
+		var clusterIndex = prefixSummedWidths.BinarySearch(
+			((float)p.X - GetAlignmentOffsetForLine(line), default),
+			Comparer<(float sumUntilAfterCluster, Cluster cluster)>.Create(
+				static (a, b) => a.sumUntilAfterCluster.CompareTo(b.sumUntilAfterCluster)));
+
+		if (clusterIndex < 0)
+		{
+			clusterIndex = ~clusterIndex;
+		}
+
+		var cluster = prefixSummedWidths[clusterIndex].cluster;
+		var right = prefixSummedWidths[clusterIndex].sumUntilAfterCluster;
+		var left = right - cluster.width;
+		var closerToLeftEdge = p.X - GetAlignmentOffsetForLine(line) - left < right - (p.X - GetAlignmentOffsetForLine(line));
+		var index = cluster.rtl
+			? closerToLeftEdge ? cluster.end : cluster.start
+			: closerToLeftEdge ? cluster.start : cluster.end;
+		return index - (ignoreEndingNewLine && line.end == index ? TrailingCRLFInLine(line) : 0);
+	}
+
+	public Hyperlink? GetHyperlinkAt(Point point)
+	{
+		var index = GetIndexAt(point, false, false);
+		if (index == -1)
+		{
+			return null;
+		}
+
+		if (_hyperlinkRanges.Count == 0)
+		{
+			return null;
+		}
+
+		var rangeIndex = _hyperlinkRanges.BinarySearch(
+			(index, 0, null!),
+			Comparer<(int start, int length, Hyperlink hyperlink)>.Create(static (a, b) => a.start.CompareTo(b.start)));
+
+		if (rangeIndex < 0)
+		{
+			rangeIndex = ~rangeIndex - 1;
+		}
+
+		if (rangeIndex < 0)
+		{
+			return null;
+		}
+
+		var range = _hyperlinkRanges[rangeIndex];
+		return index >= range.start && index < range.length ? range.hyperlink : null;
+	}
+
+	/// <param name="right">when on a word boundary, decides whether to return the left or the right word</param>
+	public (int start, int length) GetWordAt(int index, bool right)
+	{
+		if (index == 0)
+		{
+			if (_wordBoundaries.Count == 0)
+			{
+				return (0, 0);
+			}
+			else
+			{
+				return (0, _wordBoundaries[0]);
+			}
+		}
+		else if (index == _text.Length)
+		{
+			if (right)
+			{
+				return (_text.Length, 0);
+			}
+			else if (_wordBoundaries.Count == 1)
+			{
+				return (0, _text.Length);
+			}
+			else
+			{
+				return (_wordBoundaries[^2], _text.Length);
+			}
+		}
+		else
+		{
+			var prevBoundary = 0;
+			foreach (var boundary in _wordBoundaries)
+			{
+				if (index < boundary || boundary == index && !right)
+				{
+					return (prevBoundary, boundary - prevBoundary);
+				}
+				prevBoundary = boundary;
+			}
+		}
+		throw new UnreachableException();
+	}
+
+	public (int start, int length, bool firstLine, bool lastLine, int lineIndex) GetLineAt(int index)
+	{
+		if (index == 0 || _text.Length == 0 || _lines.Count == 0)
+		{
+			return (0, 0, true, _lines.Count == 1 || (_lines.Count == 0 && _endingNewLineLineHeight is not null), 0);
+		}
+		if (index >= _lines[^1].end)
+		{
+			if (_endingNewLineLineHeight is not null)
+			{
+				return (_lines[^1].end, _text.Length - _lines[^1].end, false, true, _lines.Count);
+			}
+			else
+			{
+				return (_lines[^1].start, _lines[^1].end - _lines[^1].start, _lines.Count == 1, true, _lines.Count - 1);
+			}
+		}
+
+		var lineIndex = _lines.BinarySearch(
+			new Line { end = index + 1 },
+			Comparer<Line>.Create(static (a, b) => a.end.CompareTo(b.end)));
+
+		if (lineIndex < 0)
+		{
+			lineIndex = ~lineIndex;
+		}
+
+		var line = _lines[lineIndex];
+		return (line.start, line.end - line.start, lineIndex == 0, _endingNewLineLineHeight is null && lineIndex == _lines.Count - 1, lineIndex);
+	}
+
+	public bool IsBaseDirectionRightToLeft => _rtl;
+
+	private static List<int> GetWords(string text)
+	{
+		var boundaries = new List<int>();
+		AppendBoundaries(/* Word */ 1, text, 0, boundaries);
+		var ret = new List<int> { boundaries[0] };
+		for (var index = 1; index < boundaries.Count; index++)
+		{
+			var boundary = boundaries[index];
+
+			if (boundary - ret[^1] == 1 && (char.IsPunctuation(text[boundary - 1]) || char.IsSymbol(text[boundary - 1])) && (char.IsPunctuation(text[ret[^1] - 1]) || char.IsSymbol(text[ret[^1] - 1])))
+			{
+				ret.RemoveAt(ret.Count - 1);
+			}
+			else if (Enumerable.Range(ret[^1], boundary - ret[^1]).All(c => text[c] == ' ') && !char.IsWhiteSpace(text[ret[^1] - 1]))
+			{
+				ret.RemoveAt(ret.Count - 1);
+			}
+			ret.Add(boundary);
+		}
+
+		return ret;
+	}
+
+	private int TrailingCRLFInLine(Line? line)
+	{
+		if (_text.Length == 0)
+		{
+			return 0;
+		}
+		return line is null ? TrailingCRLFCount(_text, _text.Length) : TrailingCRLFCount(_text, line.Value.end);
+	}
+
+	private float GetAlignmentOffsetForLine(Line? line)
+	{
+		var (lineWidth, lineWidthWithoutTrailingSpaces) = line is null ? (0, 0) : (line.Value.width, line.Value.widthWithoutTrailingSpaces);
+		var totalWidth = (float)_availableSize.Width;
+
+		float alignmentOffset;
+		// Trailing whitespace at the end of a line is assigned the paragraph embedding level
+		// So the trailing space is always on the left if RTL and always on the right if LTR
+		if (_rtl)
+		{
+			alignmentOffset = _textAlignment switch
+			{
+				TextAlignment.Center when lineWidthWithoutTrailingSpaces <= totalWidth => (totalWidth - lineWidthWithoutTrailingSpaces) / 2,
+				TextAlignment.Right when lineWidthWithoutTrailingSpaces <= totalWidth => totalWidth - lineWidth,
+				_ => 0
+			};
+			alignmentOffset = Math.Min(alignmentOffset, totalWidth - lineWidth);
+		}
+		else
+		{
+			alignmentOffset = _textAlignment switch
+			{
+				TextAlignment.Center when lineWidthWithoutTrailingSpaces <= totalWidth => (totalWidth - lineWidthWithoutTrailingSpaces) / 2,
+				TextAlignment.Right when lineWidthWithoutTrailingSpaces <= totalWidth => totalWidth - lineWidthWithoutTrailingSpaces,
+				_ => 0
+			};
+			alignmentOffset = Math.Max(alignmentOffset, 0);
+		}
+
+		return alignmentOffset;
 	}
 
 	private static FontDetails? GetFallbackFont(int codepoint, float fontSize, FontWeight fontWeight, FontStretch fontStretch, FontStyle fontStyle, IFontCacheUpdateListener fontListener)
@@ -683,309 +1403,8 @@ internal readonly partial struct UnicodeText : IParsedText
 		return null;
 	}
 
-	private static IEnumerable<(ReadonlyInlineCopy Inline, int startInInline, int endInInline)> GroupByInline(List<BidiRun> line)
+	private static unsafe void AppendBoundaries(int boundaryType, string text, int outputBaseOffset, List<int> list)
 	{
-		foreach (var group in line.GroupBy(r => r.inline))
-		{
-			var inline = group.Key;
-			var groupAsArray = group.ToArray();
-			var startInInline = groupAsArray[0].startInInline;
-			var endInInline = groupAsArray[^1].endInInline;
-			var i = startInInline;
-			while (inline.Text.IndexOf('\t', i) is var tabIndex && tabIndex != -1 && tabIndex < endInInline)
-			{
-				if (i != tabIndex)
-				{
-					yield return (inline, i, tabIndex);
-				}
-				yield return (inline, tabIndex, tabIndex + 1);
-				i = tabIndex + 1;
-				if (i == endInInline)
-				{
-					break;
-				}
-			}
-
-			if (i != endInInline)
-			{
-				yield return (inline, i, endInInline);
-			}
-		}
-	}
-
-	private static float MeasureContiguousRunSequence(float currentLineWidth, List<BidiRun> logicallyOrderedRuns, int firstRunIndex, int startingIndexInFirstRun, int endRunIndex, int endIndexInLastRun, bool rtl)
-	{
-		var backup1 = logicallyOrderedRuns[firstRunIndex];
-		if (endRunIndex - 1 < 0 || endRunIndex - 1 >= logicallyOrderedRuns.Count)
-		{
-		}
-		var backup2 = logicallyOrderedRuns[endRunIndex - 1];
-		logicallyOrderedRuns[firstRunIndex] = logicallyOrderedRuns[firstRunIndex] with { startInInline = startingIndexInFirstRun };
-		logicallyOrderedRuns[endRunIndex - 1] = logicallyOrderedRuns[endRunIndex - 1] with { endInInline = endIndexInLastRun };
-
-		float usedWidth = currentLineWidth;
-		for (var index = endRunIndex - 2; index >= firstRunIndex; index--)
-		{
-			var run = logicallyOrderedRuns[index];
-			var start_ = index == firstRunIndex ? startingIndexInFirstRun : run.startInInline;
-			var glyphs = ShapeRun(run.inline.Text[start_..run.endInInline], run.rtl, run.fontDetails, usedWidth, ignoreTrailingSpaces: false);
-			usedWidth += RunWidth(glyphs, run.inline.FontDetails);
-		}
-
-		var lastRun = logicallyOrderedRuns[endRunIndex - 1];
-		var start = firstRunIndex == endRunIndex - 1 ? startingIndexInFirstRun : lastRun.startInInline;
-		var glyphsWithoutTrailingSpaces = ShapeRun(lastRun.inline.Text[start..endIndexInLastRun], lastRun.rtl, lastRun.fontDetails, usedWidth, ignoreTrailingSpaces: true);
-
-		logicallyOrderedRuns[firstRunIndex] = backup1;
-		logicallyOrderedRuns[endRunIndex - 1] = backup2;
-
-		return usedWidth - currentLineWidth + RunWidth(glyphsWithoutTrailingSpaces, lastRun.inline.FontDetails);
-	}
-
-	private static IEnumerable<(int endRunIndex, int endIndexInLastRun)> SplitByLineBreakingOpportunities(List<BidiRun> logicallyOrderedRuns, List<(int indexInInline, ReadonlyInlineCopy inline)> logicallyOrderedLineBreakingOpportunities)
-	{
-		var current = 0;
-		foreach (var lineBreakingOpportunity in logicallyOrderedLineBreakingOpportunities)
-		{
-			while (logicallyOrderedRuns[current].inline != lineBreakingOpportunity.inline || logicallyOrderedRuns[current].endInInline < lineBreakingOpportunity.indexInInline)
-			{
-				current++;
-			}
-			yield return (current + 1, lineBreakingOpportunity.indexInInline);
-			if (logicallyOrderedRuns[current].endInInline == lineBreakingOpportunity.indexInInline)
-			{
-				current++;
-			}
-		}
-	}
-
-	// This is completely outside the Unicode standard. We're in a situation where a sequence of characters without
-	// any line breaking opportunities doesn't fit in a line so we need to put the part that fits on the line and
-	// the rest on the next line.
-	private static (int endRunIndex, int endIndexInLastRun) SplitRunSequenceForWrapping(float lineWidth, bool rtl, List<BidiRun> logicallyOrderedRuns, int firstRunIndex, int startingIndexInFirstRun, int endRunIndex, int endIndexInLastRun)
-	{
-		var start = (firstRunIndex + 1, startingIndexInFirstRun + 1);
-		var end = (endRunIndex, endIndexInLastRun);
-
-		var checkInvariant = ((int, int) test) =>
-		{
-			var widthWithoutTrailingSpaces = MeasureContiguousRunSequence(lineWidth, logicallyOrderedRuns, firstRunIndex, startingIndexInFirstRun, test.Item1, test.Item2, rtl);
-			return widthWithoutTrailingSpaces <= lineWidth;
-		};
-
-		var getMidpoint = ((int, int) start, (int, int) end) =>
-		{
-			if (start.Item1 == end.Item1)
-			{
-				var midIndexInLastRun = (start.Item2 + end.Item2) / 2;
-				if (char.IsHighSurrogate(logicallyOrderedRuns[start.Item1 - 1].inline.Text[midIndexInLastRun - 1]))
-				{
-					midIndexInLastRun--;
-				}
-				return (start.Item1, midIndexInLastRun);
-			}
-			else
-			{
-				var midRunIndex = (start.Item1 + end.Item1) / 2;
-				return (midRunIndex, logicallyOrderedRuns[midRunIndex].startInInline + 1);
-			}
-		};
-
-		var (endRunIndexResult, endIndexInLastRunResult) = FindMaxSatisfyingInvariant(start, end, checkInvariant, getMidpoint);
-
-		if (checkInvariant((endRunIndexResult, endIndexInLastRunResult)))
-		{
-			return (endRunIndexResult, endIndexInLastRunResult);
-		}
-
-		// nothing fits, default to having a single character in the first run
-		return (firstRunIndex + 1, startingIndexInFirstRun + (char.IsSurrogate(logicallyOrderedRuns[firstRunIndex].inline.Text[startingIndexInFirstRun]) ? 2 : 1));
-	}
-
-	private static T FindMaxSatisfyingInvariant<T>(
-			T start,
-			T end,
-			Func<T, bool> checkInvariant,
-			Func<T, T, T> getMidpoint)
-	{
-		while (!EqualityComparer<T>.Default.Equals(start, end))
-		{
-			var mid = getMidpoint(start, end);
-
-			if (EqualityComparer<T>.Default.Equals(mid, start))
-			{
-				return checkInvariant(end) ? end : start;
-			}
-
-			if (checkInvariant(mid))
-			{
-				start = mid;
-			}
-			else
-			{
-				end = mid;
-			}
-		}
-		return start;
-	}
-
-	private static List<List<BidiRun>> ApplyLineBreaking(float lineWidth, List<BidiRun> logicallyOrderedRuns,
-		List<(int endRunIndex, int endIndexInLastRun)> lineBreakingOpporunities, bool rtl,
-		TextWrapping textWrapping)
-	{
-		var lines = new List<List<BidiRun>>();
-
-		(int firstRunIndex, int startingIndexInFirstRun) = (0, 0);
-		(int endRunIndex, int endIndexInLastRun)? prevInSameLine = null;
-
-		for (var lineBreakingOpportunityIndex = 0; lineBreakingOpportunityIndex < lineBreakingOpporunities.Count; lineBreakingOpportunityIndex++)
-		{
-			var lineBreakingOpporunity = lineBreakingOpporunities[lineBreakingOpportunityIndex];
-
-			var widthWithoutTrailingSpaces = MeasureContiguousRunSequence(0, logicallyOrderedRuns, firstRunIndex, startingIndexInFirstRun, lineBreakingOpporunity.endRunIndex, lineBreakingOpporunity.endIndexInLastRun, rtl);
-
-			if (widthWithoutTrailingSpaces > lineWidth)
-			{
-				// If there was no text wrapping, then everything would fit on the same line
-				CI.Assert(textWrapping is TextWrapping.Wrap or TextWrapping.WrapWholeWords);
-
-				if (prevInSameLine is not null)
-				{
-					// after including the current run sequence on the line, things won't fit anymore
-					// so it moves to the next line regardless of whether we're in Wrap or WrapWholeWords
-					var line = new List<BidiRun>();
-					line.AddRange(logicallyOrderedRuns.Slice(firstRunIndex, prevInSameLine.Value.endRunIndex - firstRunIndex));
-					line[0] = line[0] with { startInInline = startingIndexInFirstRun };
-					line[^1] = line[^1] with { endInInline = prevInSameLine.Value.endIndexInLastRun };
-					lines.Add(line);
-					(firstRunIndex, startingIndexInFirstRun) = (prevInSameLine.Value.endRunIndex - 1, prevInSameLine.Value.endIndexInLastRun);
-					prevInSameLine = null;
-					lineBreakingOpportunityIndex--; // retry the line breaking opportunity
-				}
-				else
-				{
-					// only one "non-line-breakable" sequence is on this line but it still won't fit
-					if (textWrapping is TextWrapping.WrapWholeWords || logicallyOrderedRuns[firstRunIndex].inline.Text[startingIndexInFirstRun] == ' ')
-					{
-						// WrapWholeWords or nothing but spaces: move to the next line
-						var line = new List<BidiRun>();
-						line.AddRange(logicallyOrderedRuns.Slice(firstRunIndex, lineBreakingOpporunity.endRunIndex - firstRunIndex));
-						line[0] = line[0] with { startInInline = startingIndexInFirstRun };
-						line[^1] = line[^1] with { endInInline = lineBreakingOpporunity.endIndexInLastRun };
-						lines.Add(line);
-						(firstRunIndex, startingIndexInFirstRun) = (lineBreakingOpporunity.endRunIndex - 1, lineBreakingOpporunity.endIndexInLastRun);
-					}
-					else // Wrap
-					{
-						(int endRunIndex, int endIndexInLastRun) = SplitRunSequenceForWrapping(lineWidth, rtl, logicallyOrderedRuns, firstRunIndex, startingIndexInFirstRun, lineBreakingOpporunity.endRunIndex, lineBreakingOpporunity.endIndexInLastRun);
-						var line = new List<BidiRun>();
-						line.AddRange(logicallyOrderedRuns.Slice(firstRunIndex, endRunIndex - firstRunIndex));
-						line[0] = line[0] with { startInInline = startingIndexInFirstRun };
-						line[^1] = line[^1] with { endInInline = endIndexInLastRun };
-						lines.Add(line);
-						(firstRunIndex, startingIndexInFirstRun) = (endRunIndex - 1, endIndexInLastRun);
-						if (endRunIndex != lineBreakingOpporunity.endRunIndex || endIndexInLastRun != lineBreakingOpporunity.endIndexInLastRun)
-						{
-							lineBreakingOpportunityIndex--; // retry the line breaking opportunity
-						} // else we took the whole thing and couldn't split the sequence any further
-					}
-				}
-			}
-			else if (IsLineBreak(logicallyOrderedRuns[lineBreakingOpporunity.endRunIndex - 1].inline.Text, lineBreakingOpporunity.endIndexInLastRun))
-			{
-				var line = new List<BidiRun>();
-				line.AddRange(logicallyOrderedRuns.Slice(firstRunIndex, lineBreakingOpporunity.endRunIndex - firstRunIndex));
-				line[0] = line[0] with { startInInline = startingIndexInFirstRun };
-				line[^1] = line[^1] with { endInInline = lineBreakingOpporunity.endIndexInLastRun };
-				lines.Add(line);
-				(firstRunIndex, startingIndexInFirstRun) = (lineBreakingOpporunity.endRunIndex - 1, lineBreakingOpporunity.endIndexInLastRun);
-				prevInSameLine = null;
-			}
-			else
-			{
-				prevInSameLine = lineBreakingOpporunity;
-			}
-
-			if (firstRunIndex < logicallyOrderedRuns.Count - 1 && logicallyOrderedRuns[firstRunIndex].endInInline == startingIndexInFirstRun)
-			{
-				// The current line starts at the end of a run, so it actually means that it starts at the beginning of
-				// the next run
-				firstRunIndex++;
-				startingIndexInFirstRun = logicallyOrderedRuns[firstRunIndex].startInInline;
-			}
-		}
-
-		if (prevInSameLine is not null)
-		{
-			var line = new List<BidiRun>();
-			line.AddRange(logicallyOrderedRuns.Slice(firstRunIndex, prevInSameLine.Value.endRunIndex - firstRunIndex));
-			line[0] = line[0] with { startInInline = startingIndexInFirstRun };
-			line[^1] = line[^1] with { endInInline = prevInSameLine.Value.endIndexInLastRun };
-			(firstRunIndex, startingIndexInFirstRun) = (prevInSameLine.Value.endRunIndex - 1, prevInSameLine.Value.endIndexInLastRun);
-			lines.Add(line);
-		}
-
-		if (IsLineBreak(logicallyOrderedRuns[firstRunIndex].inline.Text, startingIndexInFirstRun))
-		{
-			var line = new List<BidiRun>();
-			lines.Add(line);
-		}
-
-#if DEBUG
-		CI.Assert(lines.Count > 0);
-		// all the lines have content except possibly the last line, because we only move to a new line when we hit a line break
-		foreach (var line in lines)
-		{
-			CI.Assert(line.Count > 0 || line == lines[^1]);
-		}
-#endif
-
-		return lines;
-	}
-
-	private static IEnumerable<int> GetLineBreakingOpportunities(string text)
-	{
-		if (OperatingSystem.IsBrowser())
-		{
-			var ret = new List<int>();
-			var breaker = ICU.BrowserICUSymbols.init_line_breaker(text);
-			if (breaker == IntPtr.Zero)
-			{
-				typeof(UnicodeText).LogError()?.Error($"Failed to create a break iterator for input text '{text}' of unicode codepoints [{string.Join(',', text.EnumerateRunes().Select(r => r.Value))}]. Falling back to naive space-based line breaking.");
-				var lines = new List<int>();
-				for (int i = 0; i < text.Length;)
-				{
-					var inc = text.Length > i + 1 && text[i] == '\r' && text[i + 1] == '\n' ? 2 : 1;
-					if (char.IsWhiteSpace(text[i]))
-					{
-						lines.Add(i + inc);
-					}
-
-					i += inc;
-				}
-
-				if (lines.Count == 0 || lines[^1] != text.Length)
-				{
-					lines.Add(text.Length);
-				}
-				return lines;
-			}
-
-			for (var next = ICU.BrowserICUSymbols.next_line_breaking_opportunity(breaker); next != -1; next = ICU.BrowserICUSymbols.next_line_breaking_opportunity(breaker))
-			{
-				ret.Add(next);
-			}
-			return ret;
-		}
-		else
-		{
-			return GetBoundaries( /* Line */ 2, text);
-		}
-	}
-
-	private static unsafe List<int> GetBoundaries(int boundaryType, string text)
-	{
-		var ret = new List<int>();
 		fixed (char* locale = &CultureInfo.CurrentUICulture.Name.GetPinnableReference())
 		{
 			fixed (char* textPtr = &text.GetPinnableReference())
@@ -995,808 +1414,10 @@ internal readonly partial struct UnicodeText : IParsedText
 				ICU.GetMethod<ICU.ubrk_first>()(breakIterator);
 				while (ICU.GetMethod<ICU.ubrk_next>()(breakIterator) is var next && next != /* UBRK_DONE */ -1)
 				{
-					ret.Add(next);
+					list.Add(next + outputBaseOffset);
 				}
 				ICU.GetMethod<ICU.ubrk_close>()(breakIterator);
 			}
-		}
-		return ret;
-	}
-
-	private static List<int> GetWords(string text)
-	{
-		var boundaries = GetBoundaries(/* Word */ 1, text);
-		var ret = new List<int> { boundaries[0] };
-		for (var index = 1; index < boundaries.Count; index++)
-		{
-			var boundary = boundaries[index];
-
-			if (boundary - ret[^1] == 1 && (char.IsPunctuation(text[boundary - 1]) || char.IsSymbol(text[boundary - 1])) && (char.IsPunctuation(text[ret[^1] - 1]) || char.IsSymbol(text[ret[^1] - 1])))
-			{
-				ret.RemoveAt(ret.Count - 1);
-			}
-			else if (Enumerable.Range(ret[^1], boundary - ret[^1]).All(c => text[c] == ' ') && !char.IsWhiteSpace(text[ret[^1] - 1]))
-			{
-				ret.RemoveAt(ret.Count - 1);
-			}
-			ret.Add(boundary);
-		}
-
-		return ret;
-	}
-
-	private static List<BidiRun> SplitTextIntoLogicallyOrderedBidiRuns(ReadonlyInlineCopy inline, IFontCacheUpdateListener fontListener)
-	{
-		using var _ = ICU.CreateBiDiAndSetPara(inline.Text, 0, inline.Text.Length, (byte)(inline.FlowDirection == FlowDirection.LeftToRight ? UBIDI_LTR : UBIDI_RTL), out var bidi);
-
-		var runCount = ICU.GetMethod<ICU.ubidi_countRuns>()(bidi, out int countRunsErrorCode);
-		ICU.CheckErrorCode<ICU.ubidi_countRuns>(countRunsErrorCode);
-
-		var logicallyOrderedRuns = new List<BidiRun>(runCount);
-		for (var runIndex = 0; runIndex < runCount; runIndex++)
-		{
-			// using bidi.GetLogicalRun instead returned weird results especially in rtl text.
-			var level = ICU.GetMethod<ICU.ubidi_getVisualRun>()(bidi, runIndex, out var logicalStart, out var length);
-			CI.Assert(level is UBIDI_LTR or UBIDI_RTL);
-
-			var currentFontDetails = inline.FontDetails;
-			var currentFontSplitStart = logicalStart;
-			for (var i = logicalStart; i < logicalStart + length; i += char.IsSurrogate(inline.Text, i) ? 2 : 1)
-			{
-				FontDetails newFontDetails;
-				if (char.ConvertToUtf32(inline.Text, i) is var codepoint && !inline.FontDetails.SKFont.ContainsGlyph(codepoint))
-				{
-					newFontDetails = GetFallbackFont(codepoint, (float)inline.FontSize, inline.FontWeight, inline.FontStretch, inline.FontStyle, fontListener) ?? inline.FontDetails;
-				}
-				else
-				{
-					newFontDetails = inline.FontDetails;
-				}
-
-				var isTab = inline.Text[i] is '\t';
-				if (newFontDetails != currentFontDetails || isTab)
-				{
-					if (currentFontSplitStart != i)
-					{
-						logicallyOrderedRuns.Add(new BidiRun(inline, currentFontSplitStart, i, level is UBIDI_RTL, currentFontDetails));
-					}
-					currentFontDetails = newFontDetails;
-					currentFontSplitStart = i;
-					if (isTab)
-					{
-						logicallyOrderedRuns.Add(new BidiRun(inline, currentFontSplitStart, i + 1, level is UBIDI_RTL, currentFontDetails));
-						currentFontSplitStart = i + 1;
-					}
-				}
-			}
-
-			if (currentFontSplitStart != logicalStart + length)
-			{
-				logicallyOrderedRuns.Add(new BidiRun(inline, currentFontSplitStart, logicalStart + length, level is UBIDI_RTL, currentFontDetails));
-			}
-		}
-
-		logicallyOrderedRuns.Sort((run, bidiRun) => run.startInInline - bidiRun.startInInline);
-
-#if DEBUG
-		CI.Assert(logicallyOrderedRuns[0].startInInline == 0);
-		for (var i = 0; i < runCount - 1; i++)
-		{
-			CI.Assert(logicallyOrderedRuns[i].endInInline != logicallyOrderedRuns[i].startInInline && logicallyOrderedRuns[i].endInInline == logicallyOrderedRuns[i + 1].startInInline);
-		}
-		CI.Assert(logicallyOrderedRuns[^1].endInInline != logicallyOrderedRuns[^1].startInInline && logicallyOrderedRuns[^1].endInInline == inline.Text.Length);
-#endif
-
-		return logicallyOrderedRuns;
-	}
-
-	/// <param name="currentLineWidth">Only used for tab stop width calculation. Null to ignore this case.</param>
-	private static (GlyphInfo info, UnoGlyphPosition position)[] ShapeRun(string textRun, bool rtl, FontDetails fontDetails, float? currentLineWidth, bool ignoreTrailingSpaces)
-	{
-		if (ignoreTrailingSpaces)
-		{
-			textRun = textRun[..^TrailingSpaceCount(textRun, 0, textRun.Length)];
-		}
-
-		CI.Assert(textRun.Length < 2 || !textRun[..^2].Contains("\r\n"));
-		using var buffer = new Buffer();
-		buffer.AddUtf16(textRun, 0, textRun is [.., '\r', '\n'] ? textRun.Length - 1 : textRun.Length);
-		buffer.GuessSegmentProperties();
-		buffer.Direction = rtl ? Direction.RightToLeft : Direction.LeftToRight;
-		// TODO: ligatures
-		fontDetails.Font.Shape(buffer, new Feature(new Tag('l', 'i', 'g', 'a'), 0));
-		var positions = buffer.GetGlyphPositionSpan();
-		var infos = buffer.GetGlyphInfoSpan();
-		var count = buffer.Length;
-		var ret = new (GlyphInfo, UnoGlyphPosition)[count];
-		for (var i = 0; i < count; i++)
-		{
-			ret[i] = (infos[i], new UnoGlyphPosition(positions[i], positions[i].XAdvance));
-		}
-
-		// Fonts will give a width > 0 to \r, so we hardcode the width here.
-		// TODO: make this cleaner somehow
-		var isTab = textRun is "\t";
-		if ((isTab || IsLineBreak(textRun, textRun.Length)) && infos[^1].Cluster == textRun.Length - (textRun is [.., '\r', '\n'] ? 2 : 1))
-		{
-			fontDetails.Font.TryGetGlyph(' ', out var spaceCodepoint);
-			var tabWidth = TabStopWidth / fontDetails.TextScale.textScaleX;
-			// 1 and not 0 to avoid issues related to newlines having zero width. This way, it's practically 0 but not == 0.
-			var xAdvance = isTab ? tabWidth - (currentLineWidth / fontDetails.TextScale.textScaleX ?? 1) % tabWidth : 1;
-			ret[^1] = (infos[^1] with { Codepoint = spaceCodepoint }, new UnoGlyphPosition(positions[^1] with { XAdvance = (int)xAdvance }, xAdvance));
-		}
-
-		if (ret is [])
-		{
-			// Even though textRun is nonempty and fontDetails contains a font that can shape all the characters in it,
-			// Font.Shape may still decide to yield 0 glyphs.
-			fontDetails.Font.TryGetGlyph(' ', out var spaceCodepoint);
-			ret = [(new GlyphInfo { Codepoint = spaceCodepoint }, new UnoGlyphPosition(new GlyphPosition(), 0))];
-		}
-		return ret;
-	}
-
-	private List<LayoutedLine> LayoutLines(List<(List<ShapedLineBrokenBidiRun> runs, int startInText, int endInText)> lines, TextAlignment textAlignment, LineStackingStrategy lineStackingStrategy, float lineHeight, float availableWidth, FontDetails defaultFontDetails)
-	{
-		var layoutedLines = new List<LayoutedLine>();
-		float currentLineY = 0;
-		for (var lineIndex = 0; lineIndex < lines.Count; lineIndex++)
-		{
-			var line = lines[lineIndex];
-			var lineRuns = lines[lineIndex].runs;
-			var layoutedRuns = new List<LayoutedLineBrokenBidiRun>(lineRuns.Count);
-			LayoutedLine layoutedLine;
-			if (lineRuns.Count == 0)
-			{
-				// Only the last line can be empty. All other lines either have a line break character or actual content since you can't wrap to a new line without having any content on the initial line.
-				CI.Assert(lineIndex == lines.Count - 1 && lineIndex != 0);
-				var (currentLineHeight, baselineOffset) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, defaultFontDetails, false, true);
-				layoutedLine = new LayoutedLine(currentLineHeight, baselineOffset, lineIndex, 0, currentLineY, line.startInText, line.endInText, new());
-			}
-			else
-			{
-				float currentLineX = 0;
-				for (var runIndex = 0; runIndex < lineRuns.Count; runIndex++)
-				{
-					var run = lineRuns[runIndex];
-					float runX = 0;
-					var glyphs = new LayoutedGlyphDetails[run.glyphs.Length];
-					var layoutedRun = new LayoutedLineBrokenBidiRun(run.Inline, run.startInInline, run.endInInline, currentLineX, default, glyphs, run.fontDetails, run.rtl, null!, runIndex);
-					layoutedRuns.Add(layoutedRun);
-					for (var i = 0; i < glyphs.Length; i++)
-					{
-						var glyph = run.glyphs[i];
-						glyphs[i] = new LayoutedGlyphDetails(glyph.info, glyph.position, runX, default, layoutedRun);
-						runX += GlyphWidth(glyph.position, run.fontDetails);
-					}
-
-					if (run.Inline.Text[run.startInInline] is '\t')
-					{
-						runX = TabStopWidth - currentLineX % TabStopWidth;
-					}
-
-					layoutedRun.width = runX;
-					currentLineX += runX;
-				}
-
-				var lineWidth = currentLineX;
-				var alignmentOffset = textAlignment switch
-				{
-					TextAlignment.Center when lineWidth <= availableWidth => (availableWidth - lineWidth) / 2,
-					TextAlignment.Right when lineWidth <= availableWidth => availableWidth - lineWidth,
-					_ => 0
-				};
-
-				var fontDetailsWithMaxHeight = lineRuns.MaxBy(r => r.fontDetails.LineHeight).fontDetails;
-				var (currentLineHeight, baselineOffset) = GetLineHeightAndBaselineOffset(lineStackingStrategy, lineHeight, fontDetailsWithMaxHeight, lineIndex == 0, lineIndex == lines.Count - 1);
-				layoutedLine = new LayoutedLine(currentLineHeight, baselineOffset, lineIndex, alignmentOffset, currentLineY, line.startInText, line.endInText, layoutedRuns);
-			}
-			layoutedLines.Add(layoutedLine);
-			layoutedRuns.ForEach(r => r.line = layoutedLine);
-			currentLineY += layoutedLine.lineHeight;
-		}
-
-		return layoutedLines;
-	}
-
-	// TODO: we're using harfbuzz clusters as the units for clustering/"atomization" but should we use Unicode's text segmentation algorithm instead?
-	// and how are they different? It seems from the HarfBuzz docs that HarfBuzz clustering by default approximates Unicode's text segmentation
-	// https://harfbuzz.github.io/working-with-harfbuzz-clusters.html
-	// https://unicode.org/reports/tr29
-	private static void CreateSourceTextFromAndToGlyphMapping(List<LayoutedLine> lines, Cluster[] textIndexToGlyphMap)
-	{
-		foreach (var line in lines)
-		{
-			foreach (var run in line.runs)
-			{
-				var runGlyphLength = run.glyphs.Length;
-				var runStart = run.startInInline + run.inline.StartIndex;
-				var runLength = run.endInInline - run.startInInline;
-
-				Cluster? previousCluster = null;
-				for (var index = run.rtl ? 0 : runGlyphLength - 1; (run.rtl && index < runGlyphLength) || (!run.rtl && index >= 0); index += run.rtl ? 1 : -1)
-				{
-					var glyphDetails = run.glyphs[index];
-					if (((run.rtl && index < runGlyphLength - 1) || (!run.rtl && index > 0)) && glyphDetails.info.Cluster == run.glyphs[index + (run.rtl ? 1 : -1)].info.Cluster)
-					{
-						continue;
-					}
-					var (startGlyphIndex, endGlyphIndex) = run.rtl
-						? (previousCluster?.glyphInRunIndexEnd ?? 0, index + 1)
-						: (index, previousCluster?.glyphInRunIndexStart ?? runGlyphLength);
-					var cluster = new Cluster(runStart + (int)glyphDetails.info.Cluster, previousCluster?.sourceTextStart ?? (runStart + runLength), run, startGlyphIndex, endGlyphIndex);
-					previousCluster = cluster;
-					for (var i = startGlyphIndex; i < endGlyphIndex; i++)
-					{
-						run.glyphs[i].cluster = cluster;
-					}
-					for (var i = cluster.sourceTextStart; i < cluster.sourceTextEnd; i++)
-					{
-						textIndexToGlyphMap[i] = cluster;
-					}
-				}
-
-				CI.Assert(run.glyphs.All(g => g.cluster is not null));
-			}
-		}
-	}
-
-	public void Draw(in Visual.PaintingSession session,
-		(int index, CompositionBrush brush, float thickness)? caret,
-		IEnumerable<TextHighlighter> highlighters)
-	{
-		var wholeTextSlicer = new RangeSlicer<(Cluster selectionClusterStart, Cluster? selectionClusterEnd, CompositionBrush? background, Brush foreground)>(0, _text.Length);
-		foreach (var highlighter in highlighters)
-		{
-			foreach (var range in highlighter.Ranges)
-			{
-				if (range.Length != 0 && range.StartIndex < _text.Length && _text.Length > 0)
-				{
-					var selectionClusterStart = _textIndexToGlyph[range.StartIndex];
-					var selectionClusterEnd = range.StartIndex + range.Length < _text.Length ? _textIndexToGlyph[range.StartIndex + range.Length] : null;
-					wholeTextSlicer.Mark(
-						selectionClusterStart.sourceTextStart,
-						selectionClusterEnd?.sourceTextStart ?? _text.Length,
-						(selectionClusterStart,
-							selectionClusterEnd,
-							highlighter.Background?.GetOrCreateCompositionBrush(Compositor.GetSharedCompositor()),
-							highlighter.Foreground ?? _blackBrush));
-				}
-			}
-		}
-
-		using var textBlobBuilder = new SKTextBlobBuilder();
-
-		for (var index = 0; index < _lines.Count; index++)
-		{
-			List<(SKFont skFont, SKTextBlobBuilder blobBuilder, float x, float y, ushort[] glyphs, SKPoint[] positions, Range glyphsAndPositionsRange, Brush brush)> drawCommands = new();
-			List<(SKPath path, float strokeThickness)> spellCheckUnderlines = new();
-
-			var line = _lines[index];
-			var currentLineX = line.xAlignmentOffset;
-			foreach (var run in line.runs)
-			{
-				var glyphs = new ushort[run.glyphs.Length];
-				var positions = new SKPoint[run.glyphs.Length];
-				for (var i = 0; i < run.glyphs.Length; i++)
-				{
-					var glyph = run.glyphs[i];
-					glyphs[i] = (ushort)glyph.info.Codepoint;
-					positions[i] = new SKPoint(glyph.xPosInRun + glyph.position.GlyphPosition.XOffset * run.fontDetails.TextScale.textScaleX, line.y + glyph.position.GlyphPosition.YOffset * run.fontDetails.TextScale.textScaleY);
-				}
-
-				var runSlicer = new RangeSlicer<(CompositionBrush? background, Brush foreground)>(0, run.glyphs.Length);
-				foreach (var highlighter in wholeTextSlicer.GetSegments())
-				{
-					if (highlighter.HasValue)
-					{
-						var (selectionLeft, selectionRight) = ClusterRangeToRunGlyphRange(run, highlighter.Value.selectionClusterStart, highlighter.Value.selectionClusterEnd);
-						runSlicer.Mark(selectionLeft, selectionRight, (highlighter.Value.background, highlighter.Value.foreground));
-					}
-				}
-
-				foreach (var slice in runSlicer.GetSegments())
-				{
-					var leftX = positions[slice.Start].X;
-					var rightX = positions[slice.End - 1].X + GlyphWidth(run.glyphs[slice.End - 1].position, run.fontDetails);
-					var selectionRect = new SKRect(currentLineX + leftX, line.y, currentLineX + rightX, line.y + line.lineHeight);
-					var (background, foreground) = slice.HasValue ? slice.Value : (null, run.inline.Foreground);
-					background?.Paint(session.Canvas, session.Opacity, selectionRect);
-					drawCommands.Add((run.fontDetails.SKFont, textBlobBuilder, currentLineX, line.baselineOffset, glyphs, positions, new Range((int)slice.Start, (int)slice.End), foreground));
-				}
-
-				var runStartIndex = run.startInInline + run.inline.StartIndex;
-				var runEndIndex = run.endInInline + run.inline.StartIndex;
-
-				if (_corrections is not null)
-				{
-					var correctionIndex = _wordBoundaries.BinarySearch(0, _corrections.Count, runStartIndex, null);
-					if (correctionIndex < 0)
-					{
-						correctionIndex = ~correctionIndex;
-					}
-					else
-					{
-						correctionIndex++;
-					}
-
-					while (correctionIndex < _corrections.Count && (correctionIndex == 0 ? 0 : _wordBoundaries[correctionIndex - 1]) is var wordStart && wordStart < runEndIndex)
-					{
-						var correction = _corrections[correctionIndex];
-
-						if (correction is not null)
-						{
-							var correctionClusterStart = _textIndexToGlyph[Math.Max(wordStart + correction.Value.correctionStart, runStartIndex)];
-							var correctionClusterEnd = wordStart + correction.Value.correctionEnd < runEndIndex ? _textIndexToGlyph[wordStart + correction.Value.correctionEnd] : null;
-							var (correctionLeft, correctionRight) = ClusterRangeToRunGlyphRange(run, correctionClusterStart, correctionClusterEnd);
-
-							if (correctionLeft < correctionRight)
-							{
-								var leftX = positions[correctionLeft].X;
-								var rightX = positions[correctionRight - 1].X + GlyphWidth(run.glyphs[correctionRight - 1].position, run.fontDetails);
-
-								var fontSize = (float)run.inline.FontSize;
-								var scale = fontSize / 12.0f;
-								var step = 4 * scale;
-								var amplitude = 2 * scale;
-								var yOffset = 2 * scale;
-
-								var p = new SKPath();
-								var y = line.y + line.baselineOffset + yOffset;
-								p.MoveTo(currentLineX + leftX, y);
-								for (float x = currentLineX + leftX; x + step < currentLineX + rightX; x += step)
-								{
-									p.LineTo(x + step / 2, y + amplitude);
-									p.LineTo(x + step, y);
-								}
-
-								spellCheckUnderlines.Add((p, scale));
-							}
-						}
-
-						correctionIndex++;
-					}
-				}
-
-				currentLineX += run.width;
-			}
-
-			foreach (var drawCommand in drawCommands)
-			{
-				DrawText(drawCommand.skFont,
-					drawCommand.blobBuilder,
-					drawCommand.x,
-					drawCommand.y,
-					drawCommand.glyphs.AsSpan(drawCommand.glyphsAndPositionsRange),
-					drawCommand.positions.AsSpan(drawCommand.glyphsAndPositionsRange),
-					session,
-					drawCommand.brush);
-			}
-
-			foreach (var (path, strokeThickness) in spellCheckUnderlines)
-			{
-				_spareSpellCheckPaint.StrokeWidth = strokeThickness;
-				session.Canvas.DrawPath(path, _spareSpellCheckPaint);
-			}
-		}
-
-		// if the caret index is out of range, this means that the parent TextBlock/TextBox updated the text and the
-		// caret position but a new UnicodeText instance has not been created yet. In that case, skip care rendering
-		// this frame and wait to be called again after measuring.
-		if (caret is { } c && caret.Value.index <= _text.Length)
-		{
-			c.brush.Paint(session.Canvas, session.Opacity, GetCaretRectForIndex(c.index, c.thickness).ToSKRect());
-		}
-	}
-
-	private static void DrawText(SKFont skFont, SKTextBlobBuilder blobBuilder, float x, float y, ReadOnlySpan<ushort> glyphs, ReadOnlySpan<SKPoint> positions, Visual.PaintingSession session, Brush brush)
-	{
-		// Ideally, we would want to get the path of the glyphs and then draw them using CompositionBrush.Paint, but this
-		// does not work for Emojis which don't have a path and instead must be drawn directly with SKCanvas.DrawText
-		blobBuilder.AddPositionedRun(glyphs, skFont, positions);
-		var blob1 = blobBuilder.Build(); // Build resets the blob builder
-		var paint = SetupPaint(brush, session.Opacity);
-		session.Canvas.DrawText(blob1, x, y, paint);
-	}
-
-	// The range ends to the left of positions[rangeRight].X. If the range should go to the end, rangeRight == positions.Length
-	private static (int rangeLeft, int rangeRight) ClusterRangeToRunGlyphRange(LayoutedLineBrokenBidiRun run, Cluster selectionClusterStart, Cluster? selectionClusterEnd)
-	{
-		int selectionLeft;
-		int selectionRight;
-
-		if (run.rtl)
-		{
-			selectionLeft = selectionClusterEnd?.layoutedRun == run
-				? selectionClusterEnd.glyphInRunIndexEnd
-				: selectionClusterEnd is not null && selectionClusterEnd.layoutedRun < run
-					? run.glyphs.Length
-					: 0;
-			selectionRight = selectionClusterStart.layoutedRun == run
-				? selectionClusterStart.glyphInRunIndexStart + 1
-				: selectionClusterStart.layoutedRun < run
-					? run.glyphs.Length
-					: 0;
-		}
-		else
-		{
-			selectionLeft = selectionClusterStart.layoutedRun == run
-				? selectionClusterStart.glyphInRunIndexStart
-				: selectionClusterStart.layoutedRun < run
-					? 0
-					: run.glyphs.Length;
-			selectionRight = selectionClusterEnd?.layoutedRun == run
-				? selectionClusterEnd.glyphInRunIndexStart
-				: selectionClusterEnd is not null && selectionClusterEnd.layoutedRun < run
-					? 0
-					: run.glyphs.Length;
-		}
-
-		if (selectionRight < selectionLeft)
-		{
-			selectionRight = selectionLeft;
-		}
-		return (selectionLeft, selectionRight);
-	}
-
-	private static SKPaint SetupPaint(Brush foreground, float opacity)
-	{
-		var paint = _spareDrawPaint;
-		paint.Reset();
-		paint.IsStroke = false;
-		paint.IsAntialias = true;
-
-		if (foreground is SolidColorBrush scb)
-		{
-			var scbColor = scb.Color;
-			paint.Color = new SKColor(
-				red: scbColor.R,
-				green: scbColor.G,
-				blue: scbColor.B,
-				alpha: (byte)(scbColor.A * scb.Opacity * opacity));
-		}
-		else if (foreground is GradientBrush gb)
-		{
-			var gbColor = gb.FallbackColorWithOpacity;
-			paint.Color = new SKColor(
-				red: gbColor.R,
-				green: gbColor.G,
-				blue: gbColor.B,
-				alpha: (byte)(gbColor.A * opacity));
-		}
-		else if (foreground is XamlCompositionBrushBase xcbb)
-		{
-			var gbColor = xcbb.FallbackColorWithOpacity;
-			paint.Color = new SKColor(
-				red: gbColor.R,
-				green: gbColor.G,
-				blue: gbColor.B,
-				alpha: (byte)(gbColor.A * opacity));
-		}
-
-		return paint;
-	}
-
-	private static float RunWidth((GlyphInfo info, UnoGlyphPosition position)[] glyphs, FontDetails details) => glyphs.Sum(g => GlyphWidth(g.position, details));
-	private static float GlyphWidth(UnoGlyphPosition position, FontDetails details) => position.XAdvance * details.TextScale.textScaleX;
-
-	public Rect GetCaretRectForIndex(int index, float caretThickness)
-	{
-		if (index == 0 && string.IsNullOrEmpty(_text))
-		{
-			var alignmentOffset = _textAlignment switch
-			{
-				TextAlignment.Left => 0,
-				TextAlignment.Center => _calculatedSize.Width / 2,
-				TextAlignment.Right => _calculatedSize.Width,
-				_ => throw new ArgumentOutOfRangeException()
-			};
-			return new Rect(alignmentOffset, 0, caretThickness, _defaultFontDetails.LineHeight);
-		}
-
-		if (index == _text.Length)
-		{
-			var lastLine = _lines[^1];
-			if (lastLine.runs.Count == 0)
-			{
-				// text ending in newline
-				return new Rect(_rtl ? lastLine.xAlignmentOffset - caretThickness : lastLine.xAlignmentOffset, lastLine.y, caretThickness, lastLine.lineHeight);
-			}
-			else
-			{
-				var lastCluster = _textIndexToGlyph[^1];
-				var lastGlyph = lastCluster.layoutedRun.glyphs[lastCluster.glyphInRunIndexEnd - 1];
-				var lastGlyphX = lastGlyph.xPosInRun + lastCluster.layoutedRun.xPosInLine + lastCluster.layoutedRun.line.xAlignmentOffset;
-				return new Rect(lastCluster.layoutedRun.rtl ? lastGlyphX : lastGlyphX + GlyphWidth(lastGlyph.position, lastCluster.layoutedRun.fontDetails) - caretThickness, lastCluster.layoutedRun.line.y, caretThickness, lastCluster.layoutedRun.line.lineHeight);
-			}
-		}
-		else
-		{
-			var cluster = _textIndexToGlyph[index];
-			var glyph = cluster.layoutedRun.glyphs[cluster.glyphInRunIndexStart];
-			var glyphX = glyph.xPosInRun + cluster.layoutedRun.xPosInLine + cluster.layoutedRun.line.xAlignmentOffset;
-			var rect = new Rect(cluster.layoutedRun.rtl ? glyphX + GlyphWidth(glyph.position, cluster.layoutedRun.fontDetails) - caretThickness : glyphX, cluster.layoutedRun.line.y, caretThickness, cluster.layoutedRun.line.lineHeight);
-
-			// When the index is set to be right after a run that runs in the direction of the base direction,
-			// and right at the start of a run that runs opposite to the base direction, the caret should be
-			// at the end of the logically previous run, not at the start of the "current" run.
-			var glyphRun = glyph.parentRun;
-			var isFirstGlyphInRun = (!glyphRun.rtl && glyphRun.glyphs[0] == glyph) || (glyphRun.rtl && glyphRun.glyphs[^1] == glyph);
-			if (isFirstGlyphInRun && glyphRun.indexInLine > 0)
-			{
-				if (_rtl && !glyphRun.rtl && glyphRun.indexInLine < glyphRun.line.runs.Count - 1 && glyphRun.line.runs[glyphRun.indexInLine + 1].rtl)
-				{
-					rect.X += glyphRun.width;
-				}
-				else if (!_rtl && glyphRun.rtl && glyphRun.indexInLine > 0 && !glyphRun.line.runs[glyphRun.indexInLine - 1].rtl)
-				{
-					rect.X -= glyphRun.width;
-				}
-			}
-
-			return rect;
-		}
-	}
-
-	public Rect GetRectForIndex(int index)
-	{
-		if (index == 0 && string.IsNullOrEmpty(_text))
-		{
-			var alignmentOffset = _textAlignment switch
-			{
-				TextAlignment.Left => 0,
-				TextAlignment.Center => _calculatedSize.Width / 2,
-				TextAlignment.Right => _calculatedSize.Width,
-				_ => throw new ArgumentOutOfRangeException()
-			};
-			return new Rect(alignmentOffset, 0, 0, _defaultFontDetails.LineHeight);
-		}
-
-		index = Math.Min(index, _text.Length);
-
-		if (index == _text.Length)
-		{
-			var lastRect = GetRectForIndex(index - 1);
-			return _rtl ?
-				new Rect(lastRect.Left, lastRect.Y, 0, lastRect.Height) :
-				new Rect(lastRect.Right, lastRect.Y, 0, lastRect.Height);
-		}
-		var cluster = _textIndexToGlyph[index];
-		var glyphs = cluster.layoutedRun.glyphs[cluster.glyphInRunIndexStart..cluster.glyphInRunIndexEnd];
-
-		var x = glyphs[0].xPosInRun + cluster.layoutedRun.xPosInLine + cluster.layoutedRun.line.xAlignmentOffset;
-		var y = cluster.layoutedRun.line.y;
-		var width = glyphs.Sum(g => GlyphWidth(g.position, cluster.layoutedRun.fontDetails));
-		var height = cluster.layoutedRun.line.lineHeight;
-		return new Rect(x, y, width, height);
-	}
-
-	public int GetIndexAt(Point p, bool ignoreEndingNewLine, bool extendedSelection) =>
-		GetIndexAndRunAt(p, ignoreEndingNewLine, extendedSelection).index;
-
-	private (int index, LayoutedLineBrokenBidiRun? run) GetIndexAndRunAt(Point p, bool ignoreEndingNewLine, bool extendedSelection)
-	{
-		if (_lines.Count == 0)
-		{
-			return extendedSelection ? (0, null) : (-1, null);
-		}
-
-		if (p.Y < 0)
-		{
-			if (extendedSelection)
-			{
-				p.Y = 0;
-			}
-			else
-			{
-				return (-1, null);
-			}
-		}
-
-		if (_calculatedSize.Height < p.Y)
-		{
-			if (extendedSelection)
-			{
-				p.Y = _calculatedSize.Height;
-			}
-			else
-			{
-				return (-1, null);
-			}
-		}
-
-		foreach (var line in _lines)
-		{
-			if (line.y > p.Y || line.y + line.lineHeight < p.Y)
-			{
-				continue;
-			}
-
-			if (line.runs.Count == 0)
-			{
-				CI.Assert(line == _lines[^1]);
-				return extendedSelection ? (0, _lines[^2].runs.MaxBy(r => r.indexInLine + r.inline.StartIndex)) : (-1, null);
-			}
-			{
-				if (line.runs[0] is var firstRun && firstRun.xPosInLine + firstRun.line.xAlignmentOffset > p.X)
-				{
-					if (!extendedSelection)
-					{
-						return (-1, null);
-					}
-					else
-					{
-						var index = _rtl
-							? firstRun.endInInline + firstRun.inline.StartIndex - (ignoreEndingNewLine ? TrailingCRLFCount(firstRun.inline.Text, firstRun.startInInline, firstRun.endInInline) : 0)
-							: firstRun.startInInline + firstRun.inline.StartIndex;
-						return (index, firstRun);
-					}
-				}
-			}
-
-			{
-				if (line.runs[^1] is var lastRun && lastRun.xPosInLine + lastRun.line.xAlignmentOffset + lastRun.width < p.X)
-				{
-					if (!extendedSelection)
-					{
-						return (-1, null);
-					}
-					else
-					{
-						var index = _rtl
-							? lastRun.inline.StartIndex + lastRun.startInInline
-							: lastRun.inline.StartIndex + lastRun.endInInline - (ignoreEndingNewLine ? TrailingCRLFCount(lastRun.inline.Text, lastRun.startInInline, lastRun.endInInline) : 0);
-						return (index, lastRun);
-					}
-				}
-			}
-
-			foreach (var run in line.runs)
-			{
-				if (run.xPosInLine + run.line.xAlignmentOffset > p.X || run.xPosInLine + run.line.xAlignmentOffset + run.width < p.X)
-				{
-					continue;
-				}
-
-				foreach (var glyph in run.glyphs)
-				{
-					var globalGlyphX = glyph.xPosInRun + run.xPosInLine + run.line.xAlignmentOffset;
-					var width = GlyphWidth(glyph.position, run.fontDetails);
-					if (globalGlyphX <= p.X && globalGlyphX + width >= p.X)
-					{
-						var closerToLeft = p.X - globalGlyphX < globalGlyphX + width - p.X;
-						var index = (closerToLeft && !run.rtl) || (!closerToLeft && run.rtl) ? glyph.cluster!.sourceTextStart : glyph.cluster!.sourceTextEnd;
-						return (index, run);
-					}
-				}
-			}
-		}
-
-		CI.Assert(false, "This should be unreachable");
-		return (-1, null);
-	}
-
-	public Hyperlink? GetHyperlinkAt(Point point)
-	{
-		var run = GetIndexAndRunAt(point, ignoreEndingNewLine: false, extendedSelection: false).run;
-		DependencyObject? parent = run?.inline.Inline;
-		while (parent is TextElement textElement)
-		{
-			if (parent is Hyperlink h)
-			{
-				return h;
-			}
-			parent = textElement.GetParent() as DependencyObject;
-		}
-
-		return null;
-	}
-
-	public (int start, int length) GetWordAt(int index, bool right)
-	{
-		if (index == 0)
-		{
-			if (_wordBoundaries.Count == 0)
-			{
-				return (0, 0);
-			}
-			else
-			{
-				return (0, _wordBoundaries[0]);
-			}
-		}
-		else if (index == _text.Length)
-		{
-			if (right)
-			{
-				return (_text.Length, 0);
-			}
-			else if (_wordBoundaries.Count == 1)
-			{
-				return (0, _text.Length);
-			}
-			else
-			{
-				return (_wordBoundaries[^2], _text.Length);
-			}
-		}
-		else
-		{
-			var prevBoundary = 0;
-			foreach (var boundary in _wordBoundaries)
-			{
-				if (index < boundary || boundary == index && !right)
-				{
-					return (prevBoundary, boundary - prevBoundary);
-				}
-				prevBoundary = boundary;
-			}
-		}
-		throw new UnreachableException();
-	}
-
-	public (int start, int length, bool firstLine, bool lastLine, int lineIndex) GetLineAt(int index)
-	{
-		if (_lines.Count == 0)
-		{
-			return (0, 0, true, true, 0);
-		}
-		foreach (var line in _lines)
-		{
-			if (line.startInText <= index && (line.endInText > index || (line.lineIndex == _lines.Count - 1 && line.endInText == index)))
-			{
-				return (line.startInText, line.endInText - line.startInText, line.lineIndex == 0, line.lineIndex == _lines.Count - 1, line.lineIndex);
-			}
-		}
-		throw new ArgumentOutOfRangeException("Given index is not within the range of text length.");
-	}
-
-	private static int TrailingSpaceCount(string str, int start, int end)
-	{
-		for (var i = end - 1; i >= start; i--)
-		{
-			if (str[i] != ' ')
-			{
-				return end - 1 - i;
-			}
-		}
-		return 0;
-	}
-
-	private static int TrailingCRLFCount(string str, int start, int end)
-	{
-		if (str is [.., '\r', '\n'])
-		{
-			return 2;
-		}
-		else if (str[end - 1] == '\r' || str[end - 1] == '\n')
-		{
-			return 1;
-		}
-		else
-		{
-			return 0;
-		}
-	}
-
-	private static bool IsLineBreak(string text, int indexAfterLineBreakOpportunity)
-	{
-		// https://www.unicode.org/standard/reports/tr13/tr13-5.html
-
-		switch (text[indexAfterLineBreakOpportunity - 1])
-		{
-			case '\u000A':
-			case '\u000B':
-			case '\u000C':
-			case '\u000D':
-			case '\u0085':
-			case '\u2028':
-			case '\u2029': // Paragraph separator (should apply paragraph formatting, i.e. paragraph spacing/indentation on new line, unlike other line
-						   // breaks - could matter if/when Paragraph.TextIndent/RichTextBlock.TextIndent is implemented (UWP/WinUI conformance to this
-						   // behavior was not tested).
-				return true;
-			default:
-				return false;
 		}
 	}
 
@@ -1836,15 +1457,42 @@ internal readonly partial struct UnicodeText : IParsedText
 		}
 	}
 
-	public (int replaceIndexStart, int replaceIndexEnd, List<string> suggestions)? GetSpellCheckSuggestions(int correctionStart, int correctionEnd)
+	private static float AdvanceToPixels(float xAdvance, FontDetails details) => xAdvance * details.TextScale.textScaleX;
+
+	private static int TrailingCRLFCount(string str, int end)
 	{
-		if (_spellCheckingService.Value is { } spellCheckingService)
+		if (str is [.., '\r', '\n'])
 		{
-			return spellCheckingService.GetSpellCheckSuggestions(_text, _wordBoundaries, correctionStart, correctionEnd);
+			return 2;
+		}
+		else if (str[end - 1] == '\r' || str[end - 1] == '\n')
+		{
+			return 1;
 		}
 		else
 		{
-			return null;
+			return 0;
+		}
+	}
+
+	private static bool IsLineBreak(string text, int indexAfterLineBreakOpportunity)
+	{
+		// https://www.unicode.org/standard/reports/tr13/tr13-5.html
+
+		switch (text[indexAfterLineBreakOpportunity - 1])
+		{
+			case '\u000A':
+			case '\u000B':
+			case '\u000C':
+			case '\u000D':
+			case '\u0085':
+			case '\u2028':
+			case '\u2029': // Paragraph separator (should apply paragraph formatting, i.e. paragraph spacing/indentation on new line, unlike other line
+						   // breaks - could matter if/when Paragraph.TextIndent/RichTextBlock.TextIndent is implemented (UWP/WinUI conformance to this
+						   // behavior was not tested).
+				return true;
+			default:
+				return false;
 		}
 	}
 }
